sys: Remove old hex editor files
