+++ plugins/builtin/CMakeLists.txt
@@ -27,8 +27,7 @@ add_library(${PROJECT_NAME} SHARED
         source/content/providers/gdb_provider.cpp
         source/content/providers/disk_provider.cpp
 
-        #source/content/views/view_hex_editor.cpp
-        source/content/views/view_hex_editor_new.cpp
+        source/content/views/view_hex_editor.cpp
         source/content/views/view_pattern_editor.cpp
         source/content/views/view_pattern_data.cpp
         source/content/views/view_hashes.cpp
