+++ plugins/builtin/include/content/views/view_hex_editor.hpp
@@ -1,96 +1,123 @@
 #pragma once
 
+#include <hex/api/content_registry.hpp>
+
 #include <hex/ui/view.hpp>
+#include <hex/helpers/concepts.hpp>
 #include <hex/helpers/encoding_file.hpp>
 
-#include <imgui_memory_editor.h>
-
-#include <list>
-#include <tuple>
-#include <random>
-#include <vector>
-
-namespace hex::prv {
-    class Provider;
-}
+#include <algorithm>
+#include <limits>
 
 namespace hex::plugin::builtin {
-
-    using SearchFunction = std::vector<std::pair<u64, u64>> (*)(prv::Provider *&provider, std::string string);
-
-    struct HighlightBlock {
-        struct Highlight {
-            color_t color;
-            std::vector<std::string> tooltips;
-        };
-
-        constexpr static size_t Size = 0x2000;
-
-        u64 base = 0x00;
-        std::array<Highlight, Size> highlight;
-    };
 
     class ViewHexEditor : public View {
     public:
         ViewHexEditor();
-        ~ViewHexEditor() override;
 
         void drawContent() override;
         void drawAlwaysVisible() override;
 
     private:
-        MemoryEditor m_memoryEditor;
-
-        std::vector<char> m_searchStringBuffer;
-        std::vector<char> m_searchHexBuffer;
-        SearchFunction m_searchFunction                      = nullptr;
-        std::vector<std::pair<u64, u64>> *m_lastSearchBuffer = nullptr;
-        bool m_searchRequested                               = false;
-
-        i64 m_lastSearchIndex = 0;
-        std::vector<std::pair<u64, u64>> m_lastStringSearch;
-        std::vector<std::pair<u64, u64>> m_lastHexSearch;
-
-        std::string m_gotoAddressInput;
-        bool m_gotoRequested = false;
-        bool m_evaluateGoto = false;
-
-        u64 m_baseAddress = 0;
-        u64 m_resizeSize  = 0;
-
-        std::vector<u8> m_dataToSave;
-        std::set<pl::Pattern *> m_highlightedPatterns;
-
-        std::string m_loaderScriptScriptPath;
-        std::string m_loaderScriptFilePath;
-
-        hex::EncodingFile m_currEncodingFile;
-        u8 m_highlightAlpha = 0x80;
-
-        std::list<HighlightBlock> m_highlights;
-
-        bool m_processingImportExport  = false;
-        bool m_advancedDecodingEnabled = false;
-
-        void drawSearchPopup();
-        void drawSearchInput(std::vector<char> *currBuffer, ImGuiInputTextFlags flags);
-        void performSearch(const char *buffer);
-        void performSearchNext();
-        void performSearchPrevious();
-        static int inputCallback(ImGuiInputTextCallbackData *data);
-
-        void drawGotoPopup();
-        void drawEditPopup();
+        constexpr static auto InvalidSelection = std::numeric_limits<u64>::max();
 
         void openFile(const std::fs::path &path);
-
-        void copyBytes() const;
-        void pasteBytes() const;
-        void copyString() const;
-
-        void registerEvents();
         void registerShortcuts();
+        void registerEvents();
         void registerMenuItems();
+
+        void drawCell(u64 address, u8 *data, size_t size, bool hovered);
+        void drawPopup();
+        void drawSelectionFrame(u32 x, u32 y, u64 byteAddress, u16 bytesPerCell, const ImVec2 &cellPos, const ImVec2 &cellSize);
+
+    public:
+        void setSelection(const Region &region) { this->setSelection(region.getStartAddress(), region.getEndAddress()); }
+        void setSelection(i128 start, i128 end) {
+            if (!ImHexApi::Provider::isValid()) return;
+
+            const size_t maxAddress = ImHexApi::Provider::get()->getActualSize() - 1;
+
+            this->m_selectionChanged = this->m_selectionStart != start || this->m_selectionEnd != end;
+
+            this->m_selectionStart = std::clamp<decltype(start)>(start, 0, maxAddress);
+            this->m_selectionEnd = std::clamp<decltype(end)>(end, 0, maxAddress);
+
+            if (this->m_selectionChanged) {
+                EventManager::post<EventRegionSelected>(this->getSelection());
+            }
+        }
+
+        [[nodiscard]] Region getSelection() const {
+            const auto start = std::min(this->m_selectionStart, this->m_selectionEnd);
+            const auto end   = std::max(this->m_selectionStart, this->m_selectionEnd);
+            const size_t size = end - start + 1;
+
+            return { start, size };
+        }
+
+        [[nodiscard]] bool isSelectionValid() const {
+            return this->m_selectionStart != InvalidSelection && this->m_selectionEnd != InvalidSelection;
+        }
+
+        void jumpToSelection() {
+            this->m_shouldScrollToSelection = true;
+        }
+
+    public:
+        class Popup {
+        public:
+            virtual ~Popup() = default;
+            virtual void draw(ViewHexEditor *editor) = 0;
+        };
+
+        [[nodiscard]] bool isAnyPopupOpen() const {
+            return this->m_currPopup != nullptr;
+        }
+
+        template<std::derived_from<Popup> T>
+        [[nodiscard]] bool isPopupOpen() const {
+            return dynamic_cast<T*>(this->m_currPopup.get()) != nullptr;
+        }
+
+        template<std::derived_from<Popup> T, typename ... Args>
+        void openPopup(Args && ...args) {
+            this->m_currPopup = std::make_unique<T>(std::forward<Args>(args)...);
+            this->m_shouldOpenPopup = true;
+        }
+
+        void closePopup() {
+            this->m_currPopup.reset();
+        }
+
+    private:
+        void drawEditMenu();
+
+    private:
+        u16 m_bytesPerRow = 16;
+
+        ContentRegistry::HexEditor::DataVisualizer *m_currDataVisualizer;
+
+        bool m_shouldScrollToSelection = false;
+        bool m_selectionChanged = false;
+        u64 m_selectionStart = InvalidSelection;
+        u64 m_selectionEnd = InvalidSelection;
+
+        u16 m_visibleRowCount = 0;
+
+        std::optional<u64> m_editingAddress;
+        bool m_shouldModifyValue = false;
+        bool m_enteredEditingMode = false;
+        std::vector<u8> m_editingBytes;
+
+        u8 m_highlightAlpha = 0x80;
+        bool m_upperCaseHex = true;
+        bool m_grayOutZero = true;
+        bool m_showAscii = true;
+
+        bool m_shouldOpenPopup = false;
+        std::unique_ptr<Popup> m_currPopup;
+
+        std::optional<EncodingFile> m_currCustomEncoding;
     };
 
 }
