+++ plugins/builtin/source/content/views/view_hex_editor.cpp
@@ -1,267 +1,941 @@
 #include "content/views/view_hex_editor.hpp"
 
-#include <hex/api/imhex_api.hpp>
 #include <hex/api/content_registry.hpp>
-
-#include <hex/providers/provider.hpp>
-#include <hex/helpers/crypto.hpp>
-#include <hex/helpers/file.hpp>
-#include <hex/helpers/fs.hpp>
-#include <hex/helpers/patches.hpp>
 #include <hex/helpers/project_file_handler.hpp>
-#include <hex/helpers/loader_script_handler.hpp>
-
-#include <content/providers/file_provider.hpp>
+#include <hex/helpers/utils.hpp>
+#include <hex/providers/buffered_reader.hpp>
+#include <hex/helpers/crypto.hpp>
 
 #include "math_evaluator.hpp"
 
-#include <GLFW/glfw3.h>
-
+#include <imgui_internal.h>
 #include <nlohmann/json.hpp>
 
 #include <thread>
-#include <filesystem>
 
 namespace hex::plugin::builtin {
 
-    ViewHexEditor::ViewHexEditor() : View("hex.builtin.view.hex_editor.name"_lang) {
+    class PopupGoto : public ViewHexEditor::Popup {
+    public:
 
-        this->m_searchStringBuffer = std::vector<char>(0xFFF, 0x00);
-        this->m_searchHexBuffer = std::vector<char>(0xFFF, 0x00);
+        void draw(ViewHexEditor *editor) override {
+            ImGui::TextUnformatted("Goto...");
+            if (ImGui::BeginTabBar("##goto_tabs")) {
+                if (ImGui::BeginTabItem("Absolute")) {
+                    this->m_mode = Mode::Absolute;
+                    ImGui::EndTabItem();
+                }
 
-        ContentRegistry::FileHandler::add({ ".hexproj" }, [](const auto &path) {
-            return ProjectFile::load(path);
+                ImGui::BeginDisabled(!editor->isSelectionValid());
+                if (ImGui::BeginTabItem("Relative")) {
+                    this->m_mode = Mode::Relative;
+                    ImGui::EndTabItem();
+                }
+                ImGui::EndDisabled();
+
+                if (ImGui::BeginTabItem("End")) {
+                    this->m_mode = Mode::End;
+                    ImGui::EndTabItem();
+                }
+
+                ImGui::SetKeyboardFocusHere();
+                ImGui::CaptureKeyboardFromApp(true);
+                if (ImGui::InputText("##input", this->m_input, ImGuiInputTextFlags_EnterReturnsTrue | ImGuiInputTextFlags_AutoSelectAll)) {
+                    if (auto result = this->m_evaluator.evaluate(this->m_input); result.has_value()) {
+                        const auto inputResult = result.value();
+                        u64 newAddress = 0x00;
+
+                        switch (this->m_mode) {
+                            case Mode::Absolute: {
+                                    newAddress = inputResult;
+                                }
+                                break;
+                            case Mode::Relative: {
+                                    const auto selection = editor->getSelection();
+                                    newAddress = selection.getStartAddress() + inputResult;
+                                }
+                                break;
+                            case Mode::End: {
+                                    newAddress = ImHexApi::Provider::get()->getActualSize() - inputResult;
+                                }
+                                break;
+                        }
+
+                        editor->setSelection(newAddress, newAddress);
+                        editor->jumpToSelection();
+                    }
+                }
+
+                ImGui::EndTabBar();
+            }
+        }
+
+    private:
+        enum class Mode : u8 {
+            Absolute,
+            Relative,
+            End
+        };
+
+        Mode m_mode = Mode::Absolute;
+
+        std::string m_input;
+        MathEvaluator<i128> m_evaluator;
+    };
+
+    class PopupFind : public ViewHexEditor::Popup {
+    public:
+        void draw(ViewHexEditor *editor) override {
+            const auto ButtonSize = ImVec2(ImGui::CalcTextSize(ICON_VS_SEARCH).x, ImGui::GetTextLineHeight()) + ImGui::GetStyle().CellPadding * 2;
+            const auto ButtonColor = ImGui::GetStyleColorVec4(ImGuiCol_Text);
+
+            std::vector<u8> searchSequence;
+            bool shouldSearch = false;
+            bool backwards = false;
+
+            ImGui::TextUnformatted("Find...");
+            if (ImGui::BeginTabBar("##find_tabs")) {
+                if (ImGui::BeginTabItem("Hex")) {
+                    ImGui::SetKeyboardFocusHere();
+                    ImGui::CaptureKeyboardFromApp(true);
+
+                    if (ImGui::InputText("##input", this->m_input, ImGuiInputTextFlags_EnterReturnsTrue | ImGuiInputTextFlags_AutoSelectAll | ImGuiInputTextFlags_CharsHexadecimal)) {
+                        shouldSearch = true;
+                        backwards = false;
+                    }
+
+                    ImGui::SameLine();
+                    if (ImGui::IconButton(ICON_VS_SEARCH, ButtonColor, ButtonSize)) {
+                        shouldSearch = true;
+                        backwards = false;
+                        this->m_lastFind.reset();
+                    }
+
+                    if (ImGui::IconButton(ICON_VS_ARROW_UP, ButtonColor, ButtonSize)) {
+                        shouldSearch = true;
+                        backwards = true;
+                    }
+                    ImGui::SameLine();
+                    if (ImGui::IconButton(ICON_VS_ARROW_DOWN, ButtonColor, ButtonSize)) {
+                        shouldSearch = true;
+                        backwards = false;
+                    }
+
+                    if (shouldSearch) {
+                        searchSequence = decodeByteString(this->m_input);
+                    }
+
+                    ImGui::EndTabItem();
+                }
+
+                if (ImGui::BeginTabItem("String")) {
+                    ImGui::SetKeyboardFocusHere();
+                    ImGui::CaptureKeyboardFromApp(true);
+
+                    if (ImGui::InputText("##input", this->m_input, ImGuiInputTextFlags_EnterReturnsTrue | ImGuiInputTextFlags_AutoSelectAll)) {
+                        shouldSearch = false;
+                    }
+
+                    ImGui::SameLine();
+                    if (ImGui::IconButton(ICON_VS_SEARCH, ButtonColor, ButtonSize)) {
+                        shouldSearch = true;
+                        backwards = false;
+                        this->m_lastFind.reset();
+                    }
+
+                    if (ImGui::IconButton(ICON_VS_ARROW_UP, ButtonColor, ButtonSize)) {
+                        shouldSearch = true;
+                        backwards = true;
+                    }
+                    ImGui::SameLine();
+                    if (ImGui::IconButton(ICON_VS_ARROW_DOWN, ButtonColor, ButtonSize)) {
+                        shouldSearch = true;
+                        backwards = false;
+                    }
+
+                    if (shouldSearch) {
+                        searchSequence.clear();
+                        std::copy(this->m_input.begin(), this->m_input.end(), std::back_inserter(searchSequence));
+
+                        // Remove null termination
+                        searchSequence.pop_back();
+                    }
+
+                    ImGui::EndTabItem();
+                }
+
+                ImGui::EndTabBar();
+            }
+
+
+            if (!searchSequence.empty() && shouldSearch) {
+                if (auto region = this->findSequence(searchSequence, backwards); region.has_value()) {
+                    editor->setSelection(region->getStartAddress(), region->getEndAddress());
+                    editor->jumpToSelection();
+                }
+            }
+        }
+
+        std::optional<Region> findSequence(const std::vector<u8> &sequence, bool backwards) {
+            hex::prv::BufferedReader reader(ImHexApi::Provider::get());
+
+            reader.seek(this->m_lastFind.value_or(0x00));
+
+            constexpr static auto searchFunction = [](const auto &haystackBegin, const auto &haystackEnd, const auto &needleBegin, const auto &needleEnd) {
+                #if defined(OS_MACOS)
+                    return std::search(haystackBegin, haystackEnd, needleBegin, needleEnd);
+                #else
+                    return std::search(haystackBegin, haystackEnd, std::boyer_moore_horspool_searcher(needleBegin, needleEnd));
+                #endif
+            };
+
+            if (!backwards) {
+                auto occurrence = searchFunction(reader.begin(), reader.end(), sequence.begin(), sequence.end());
+                if (occurrence != reader.end()) {
+                    this->m_lastFind = occurrence.getAddress();
+                    return Region { *this->m_lastFind, *this->m_lastFind + sequence.size() - 1 };
+                }
+            } else {
+                auto occurrence = searchFunction(reader.rbegin(), reader.rend(), sequence.begin(), sequence.end());
+                if (occurrence != reader.rend()) {
+                    this->m_lastFind = occurrence.getAddress();
+                    return Region { *this->m_lastFind, *this->m_lastFind + sequence.size() - 1 };
+                }
+            }
+
+
+            return std::nullopt;
+        }
+
+        std::string m_input;
+        std::optional<u64> m_lastFind;
+    };
+
+    class PopupBaseAddress : public ViewHexEditor::Popup {
+    public:
+        explicit PopupBaseAddress(u64 baseAddress) : m_baseAddress(baseAddress) {
+
+        }
+
+        void draw(ViewHexEditor *editor) override {
+            ImGui::TextUnformatted("Base Address");
+
+            if (ImGui::InputHexadecimal("##base_address", &this->m_baseAddress, ImGuiInputTextFlags_EnterReturnsTrue)) {
+                this->setBaseAddress(this->m_baseAddress);
+                editor->closePopup();
+            }
+
+            View::confirmButtons("hex.builtin.common.set"_lang, "hex.builtin.common.cancel"_lang,
+                [&, this]{
+                    this->setBaseAddress(this->m_baseAddress);
+                    editor->closePopup();
+                },
+                [&]{
+                    editor->closePopup();
+                });
+        }
+
+    private:
+        static void setBaseAddress(u64 baseAddress) {
+            if (ImHexApi::Provider::isValid())
+                ImHexApi::Provider::get()->setBaseAddress(baseAddress);
+        }
+
+    private:
+        u64 m_baseAddress;
+    };
+
+    class PopupResize : public ViewHexEditor::Popup {
+    public:
+        explicit PopupResize(u64 currSize) : m_size(currSize) {}
+
+        void draw(ViewHexEditor *editor) override {
+            ImGui::TextUnformatted("Resize");
+
+            if (ImGui::InputHexadecimal("##resize", &this->m_size, ImGuiInputTextFlags_EnterReturnsTrue)) {
+                this->resize(static_cast<size_t>(this->m_size));
+            }
+
+            View::confirmButtons("hex.builtin.common.set"_lang, "hex.builtin.common.cancel"_lang,
+                [&, this]{
+                    this->resize(static_cast<size_t>(this->m_size));
+                    editor->closePopup();
+                },
+                [&]{
+                    editor->closePopup();
+                });
+        }
+
+    private:
+        static void resize(size_t newSize) {
+            if (ImHexApi::Provider::isValid())
+                ImHexApi::Provider::get()->resize(newSize);
+        }
+
+    private:
+        u64 m_size;
+    };
+
+    class PopupInsert : public ViewHexEditor::Popup {
+    public:
+        PopupInsert(u64 address, size_t size) : m_address(address), m_size(size) {}
+
+        void draw(ViewHexEditor *editor) override {
+            ImGui::TextUnformatted("Insert Bytes");
+
+            if (ImGui::InputHexadecimal("Address", &this->m_address, ImGuiInputTextFlags_EnterReturnsTrue)) {
+                this->insert(this->m_address, static_cast<size_t>(this->m_size));
+            }
+            if (ImGui::InputHexadecimal("Size", &this->m_size, ImGuiInputTextFlags_EnterReturnsTrue)) {
+                this->insert(this->m_address, static_cast<size_t>(this->m_size));
+            }
+
+            View::confirmButtons("hex.builtin.common.set"_lang, "hex.builtin.common.cancel"_lang,
+                [&, this]{
+                    this->insert(this->m_address, static_cast<size_t>(this->m_size));
+                    editor->closePopup();
+                },
+                [&]{
+                    editor->closePopup();
+                });
+        }
+
+    private:
+        static void insert(u64 address, size_t size) {
+            if (ImHexApi::Provider::isValid())
+                ImHexApi::Provider::get()->insert(address, size);
+        }
+
+    private:
+        u64 m_address;
+        u64 m_size;
+    };
+
+    ViewHexEditor::ViewHexEditor() : View("Hex Editor New") {
+        this->m_currDataVisualizer = ContentRegistry::HexEditor::impl::getVisualizers()["hex.builtin.visualizer.hexadecimal.8bit"];
+
+        this->registerShortcuts();
+        this->registerEvents();
+        this->registerMenuItems();
+
+        ImHexApi::HexEditor::addForegroundHighlightingProvider([this](u64 address, const u8 *data, size_t size) -> std::optional<color_t> {
+            hex::unused(address);
+
+            if (!this->m_grayOutZero)
+                return std::nullopt;
+
+            for (u32 i = 0; i < size; i++)
+                if (data[i] != 0x00)
+                    return std::nullopt;
+
+            return ImGui::GetColorU32(ImGuiCol_TextDisabled);
         });
 
-        ContentRegistry::FileHandler::add({ ".tbl" }, [this](const auto &path) {
-            this->m_currEncodingFile = EncodingFile(EncodingFile::Type::Thingy, path);
-
-            return true;
-        });
-
-        this->m_memoryEditor.ReadFn = [](const ImU8 *, size_t off) -> ImU8 {
-            auto provider = ImHexApi::Provider::get();
-            if (!provider->isAvailable() || !provider->isReadable())
-                return 0x00;
-
-            ImU8 byte;
-            provider->read(off + provider->getBaseAddress() + provider->getCurrentPageAddress(), &byte, sizeof(ImU8));
-
-            return byte;
-        };
-
-        this->m_memoryEditor.WriteFn = [](ImU8 *, size_t off, ImU8 d) -> void {
-            auto provider = ImHexApi::Provider::get();
-            if (!provider->isAvailable() || !provider->isWritable())
-                return;
-
-            provider->write(off + provider->getBaseAddress() + provider->getCurrentPageAddress(), &d, sizeof(ImU8));
-            EventManager::post<EventDataChanged>();
-            ProjectFile::markDirty();
-        };
-
-        this->m_memoryEditor.HighlightFn = [](const ImU8 *data, size_t off, bool next) -> bool {
-            bool firstByte = off == 0x00;
-
-            auto _this    = (ViewHexEditor *)(data);
-            auto provider = ImHexApi::Provider::get();
-            off += provider->getBaseAddress() + provider->getCurrentPageAddress();
-
-            u32 alpha = static_cast<u32>(_this->m_highlightAlpha) << 24;
-
-            std::optional<color_t> currColor, prevColor;
-
-            for (auto &highlightBlock : _this->m_highlights) {
-                if (off >= highlightBlock.base && off < (highlightBlock.base + HighlightBlock::Size)) {
-                    currColor = highlightBlock.highlight[off - highlightBlock.base].color;
-                }
-                if ((off - 1) >= highlightBlock.base && (off - 1) < (highlightBlock.base + HighlightBlock::Size)) {
-                    prevColor = highlightBlock.highlight[(off - 1) - highlightBlock.base].color;
-                }
-
-                if (currColor && prevColor) break;
-
-                if (firstByte)
-                    prevColor = 0x00;
-            }
-
-            if (!currColor || !prevColor) {
-                if (_this->m_highlights.size() > 0x100)
-                    _this->m_highlights.pop_front();
-
-                auto blockStartOffset   = off - (off % HighlightBlock::Size);
-                HighlightBlock newBlock = {
-                    blockStartOffset,
-                    {}
-                };
-
-                for (u32 i = 0; i < HighlightBlock::Size; i++) {
-                    std::optional<color_t> highlightColor;
-                    std::string highlightTooltip;
-
-                    for (const auto &[id, highlight] : ImHexApi::HexEditor::impl::getHighlights()) {
-                        auto &region  = highlight.getRegion();
-                        auto &color   = highlight.getColor();
-                        auto &tooltip = highlight.getTooltip();
-
-                        if (blockStartOffset + i >= region.address && blockStartOffset + i < (region.address + region.size)) {
-                            highlightColor   = (color & 0x00FFFFFF) | alpha;
-                            highlightTooltip = tooltip;
-                        }
-                    }
-
-                    for (const auto &[id, function] : ImHexApi::HexEditor::impl::getHighlightingFunctions()) {
-                        auto highlight = function(blockStartOffset + i);
-                        if (highlight.has_value()) {
-                            highlightColor   = highlightColor.has_value() ? ImAlphaBlendColors(highlight->getColor(), highlightColor.value()) : highlight->getColor();
-                            highlightTooltip = highlight->getTooltip();
-                        }
-                    }
-
-                    auto &currHighlight = newBlock.highlight[i];
-                    currHighlight.color = highlightColor.value_or(0x00);
-                    if (!highlightTooltip.empty())
-                        currHighlight.tooltips.push_back(highlightTooltip);
-                }
-
-                _this->m_highlights.push_back(std::move(newBlock));
-            }
-
-            if (next && prevColor != currColor) {
-                return false;
-            }
-
-            if (currColor.has_value() && (currColor.value() & 0x00FFFFFF) != 0x00) {
-                _this->m_memoryEditor.HighlightColor = (currColor.value() & 0x00FFFFFF) | alpha;
-                return true;
-            }
-
-            _this->m_memoryEditor.HighlightColor = 0x60C08080;
-            return false;
-        };
-
-        this->m_memoryEditor.HoverFn = [](const ImU8 *data, size_t off) {
-            auto _this = (ViewHexEditor *)(data);
-
-            bool tooltipShown = false;
-
-            auto provider = ImHexApi::Provider::get();
-            off += provider->getBaseAddress() + provider->getCurrentPageAddress();
-
-            for (auto &highlightBlock : _this->m_highlights) {
-                if (off >= highlightBlock.base && off < (highlightBlock.base + HighlightBlock::Size)) {
-                    auto &highlight = highlightBlock.highlight[off - highlightBlock.base];
-                    if (!tooltipShown && !highlight.tooltips.empty()) {
-                        ImGui::BeginTooltip();
-                        tooltipShown = true;
-                    }
-
-                    for (const auto &tooltip : highlight.tooltips) {
-                        ImGui::ColorButton(tooltip.c_str(), ImColor(highlight.color).Value);
-                        ImGui::SameLine(0, 10);
-                        ImGui::TextUnformatted(tooltip.c_str());
-                    }
-                }
-            }
-
-            if (tooltipShown)
-                ImGui::EndTooltip();
-        };
-
-        this->m_memoryEditor.DecodeFn = [](const ImU8 *data, size_t addr) -> MemoryEditor::DecodeData {
-            auto *_this = (ViewHexEditor *)data;
-
-            if (_this->m_currEncodingFile.getLongestSequence() == 0)
-                return { ".", 1, 0xFFFF8000 };
-
-            auto provider = ImHexApi::Provider::get();
-            size_t size   = std::min<size_t>(_this->m_currEncodingFile.getLongestSequence(), provider->getActualSize() - addr);
-
-            std::vector<u8> buffer(size);
-            provider->read(addr + provider->getBaseAddress() + provider->getCurrentPageAddress(), buffer.data(), size);
-
-            auto [decoded, advance] = _this->m_currEncodingFile.getEncodingFor(buffer);
-
-            ImColor color;
-            if (decoded.length() == 1 && std::isalnum(decoded[0])) color = 0xFFFF8000;
-            else if (decoded.length() == 1 && advance == 1) color = 0xFF0000FF;
-            else if (decoded.length() > 1 && advance == 1) color = 0xFF00FFFF;
-            else if (advance > 1) color = 0xFFFFFFFF;
-            else color = 0xFFFF8000;
-
-            return { std::string(decoded), advance, color };
-        };
-
-        registerEvents();
-        registerShortcuts();
-        registerMenuItems();
     }
 
-    ViewHexEditor::~ViewHexEditor() {
-        EventManager::unsubscribe<RequestOpenFile>(this);
-        EventManager::unsubscribe<RequestSelectionChange>(this);
-        EventManager::unsubscribe<EventProjectFileLoad>(this);
-        EventManager::unsubscribe<EventWindowClosing>(this);
-        EventManager::unsubscribe<RequestOpenWindow>(this);
-        EventManager::unsubscribe<EventSettingsChanged>(this);
-        EventManager::unsubscribe<EventHighlightingChanged>(this);
-        EventManager::unsubscribe<EventProviderChanged>(this);
+    constexpr static u16 getByteColumnSeparatorCount(u16 columnCount) {
+        return (columnCount - 1) / 8;
+    }
+
+    constexpr static bool isColumnSeparatorColumn(u16 currColumn, u16 columnCount) {
+        return currColumn > 0 && (currColumn) < columnCount && ((currColumn) % 8) == 0;
+    }
+
+    void ViewHexEditor::drawCell(u64 address, u8 *data, size_t size, bool hovered) {
+        if (this->m_editingAddress != address) {
+            this->m_currDataVisualizer->draw(address, data, size, this->m_upperCaseHex);
+
+            if (hovered) {
+                // Enter editing mode when double-clicking a cell
+                if (ImGui::IsMouseDoubleClicked(ImGuiMouseButton_Left)) {
+                    this->m_editingAddress = address;
+                    this->m_shouldModifyValue = false;
+                    this->m_enteredEditingMode = true;
+
+                    this->m_editingBytes.resize(size);
+                    std::memcpy(this->m_editingBytes.data(), data, size);
+                }
+            }
+        }
+        else {
+            ImGui::SetKeyboardFocusHere();
+            ImGui::CaptureKeyboardFromApp(true);
+
+            if (this->m_currDataVisualizer->drawEditing(address, this->m_editingBytes.data(), this->m_editingBytes.size(), this->m_upperCaseHex, this->m_enteredEditingMode) || this->m_shouldModifyValue) {
+
+                auto provider = ImHexApi::Provider::get();
+
+                provider->write(address, this->m_editingBytes.data(), this->m_editingBytes.size());
+
+                if (!this->m_selectionChanged && !ImGui::IsMouseDown(ImGuiMouseButton_Left) && !ImGui::IsMouseClicked(ImGuiMouseButton_Left)) {
+                    auto nextEditingAddress = *this->m_editingAddress + this->m_currDataVisualizer->getBytesPerCell();
+                    this->setSelection(nextEditingAddress, nextEditingAddress);
+
+                    if (nextEditingAddress >= provider->getSize())
+                        this->m_editingAddress = std::nullopt;
+                    else
+                        this->m_editingAddress = nextEditingAddress;
+                } else {
+                    this->m_editingAddress = std::nullopt;
+                }
+
+                this->m_shouldModifyValue = false;
+            }
+
+            this->m_enteredEditingMode = false;
+        }
+    }
+
+    static std::optional<color_t> queryBackgroundColor(u64 address, const u8 *data, size_t size) {
+        for (const auto &[id, callback] : ImHexApi::HexEditor::impl::getBackgroundHighlightingFunctions()) {
+            if (auto color = callback(address, data, size); color.has_value())
+                return color.value();
+        }
+
+        for (const auto &[id, highlighting] : ImHexApi::HexEditor::impl::getBackgroundHighlights()) {
+            if (highlighting.getRegion().overlaps({ address, size }))
+                return highlighting.getColor();
+        }
+
+        return std::nullopt;
+    }
+
+    static std::optional<color_t> queryForegroundColor(u64 address, const u8 *data, size_t size) {
+        for (const auto &[id, callback] : ImHexApi::HexEditor::impl::getForegroundHighlightingFunctions()) {
+            if (auto color = callback(address, data, size); color.has_value())
+                return color.value();
+        }
+
+        for (const auto &[id, highlighting] : ImHexApi::HexEditor::impl::getForegroundHighlights()) {
+            if (highlighting.getRegion().overlaps({ address, size }))
+                return highlighting.getColor();
+        }
+
+        return std::nullopt;
+    }
+
+    struct CustomEncodingData {
+        std::string displayValue;
+        size_t advance;
+        ImColor color;
+    };
+
+    static CustomEncodingData queryCustomEncodingData(const EncodingFile &encodingFile, u64 address) {
+        const auto longestSequence = encodingFile.getLongestSequence();
+
+        if (longestSequence == 0)
+            return { ".", 1, 0xFFFF8000 };
+
+        auto provider = ImHexApi::Provider::get();
+        size_t size   = std::min<size_t>(longestSequence, provider->getActualSize() - address);
+
+        std::vector<u8> buffer(size);
+        provider->read(address + provider->getBaseAddress() + provider->getCurrentPageAddress(), buffer.data(), size);
+
+        auto [decoded, advance] = encodingFile.getEncodingFor(buffer);
+
+        ImColor color;
+        if (decoded.length() == 1 && std::isalnum(decoded[0])) color = 0xFFFF8000;
+        else if (decoded.length() == 1 && advance == 1) color = 0xFF0000FF;
+        else if (decoded.length() > 1 && advance == 1) color = 0xFF00FFFF;
+        else if (advance > 1) color = 0xFFFFFFFF;
+        else color = 0xFFFF8000;
+
+        return { std::string(decoded), advance, color };
+    }
+
+    static void drawTooltip(u64 address, const u8 *data, size_t size) {
+        ImGui::PushStyleVar(ImGuiStyleVar_WindowPadding, scaled(ImVec2(5, 5)));
+
+        for (const auto &[id, callback] : ImHexApi::HexEditor::impl::getTooltipFunctions()) {
+            callback(address, data, size);
+        }
+
+        const auto &tooltips = ImHexApi::HexEditor::impl::getTooltips();
+        if (!tooltips.empty()) {
+            ImGui::BeginTooltip();
+
+            for (const auto &[id, tooltip] : tooltips) {
+                if (ImGui::BeginTable("##tooltips", 1, ImGuiTableFlags_RowBg | ImGuiTableFlags_NoClip)) {
+                    ImGui::TableNextRow();
+                    ImGui::TableNextColumn();
+
+                    if (tooltip.getRegion().overlaps({ address, size })) {
+                        ImGui::ColorButton(tooltip.getValue().c_str(), ImColor(tooltip.getColor()));
+                        ImGui::SameLine(0, 10);
+                        ImGui::TextUnformatted(tooltip.getValue().c_str());
+                    }
+
+                    ImGui::PushStyleColor(ImGuiCol_TableRowBg, tooltip.getColor());
+                    ImGui::PushStyleColor(ImGuiCol_TableRowBgAlt, tooltip.getColor());
+                    ImGui::EndTable();
+                    ImGui::PopStyleColor(2);
+                }
+            }
+
+            ImGui::EndTooltip();
+        }
+
+        ImGui::PopStyleVar();
+    }
+
+    void ViewHexEditor::drawSelectionFrame(u32 x, u32 y, u64 byteAddress, u16 bytesPerCell, const ImVec2 &cellPos, const ImVec2 &cellSize) {
+        if (!this->isSelectionValid()) return;
+
+        const auto selection = this->getSelection();
+        if (!Region { byteAddress, 1 }.isWithin(selection))
+            return;
+
+        const color_t SelectionFrameColor = ImGui::GetColorU32(ImGuiCol_Text);
+
+        auto drawList = ImGui::GetWindowDrawList();
+
+        // Draw vertical line at the left of first byte and the start of the line
+        if (x == 0 || byteAddress == selection.getStartAddress())
+            drawList->AddLine(cellPos, cellPos + ImVec2(0, cellSize.y), ImColor(SelectionFrameColor), 1.0F);
+
+        // Draw vertical line at the right of the last byte and the end of the line
+        if (x == u16((this->m_bytesPerRow / bytesPerCell) - 1) || (byteAddress + this->m_currDataVisualizer->getBytesPerCell()) > selection.getEndAddress())
+            drawList->AddLine(cellPos + ImVec2(cellSize.x, -1), cellPos + cellSize, ImColor(SelectionFrameColor), 1.0F);
+
+        // Draw horizontal line at the top of the bytes
+        if (y == 0 || (byteAddress - this->m_bytesPerRow) < selection.getStartAddress())
+            drawList->AddLine(cellPos, cellPos + ImVec2(cellSize.x + 1, 0), ImColor(SelectionFrameColor), 1.0F);
+
+        // Draw horizontal line at the bottom of the bytes
+        if ((byteAddress + this->m_bytesPerRow) > selection.getEndAddress())
+            drawList->AddLine(cellPos + ImVec2(0, cellSize.y), cellPos + cellSize + ImVec2(1, 0), ImColor(SelectionFrameColor), 1.0F);
+    }
+
+    void ViewHexEditor::drawPopup() {
+        // Popup windows
+        if (this->m_shouldOpenPopup) {
+            this->m_shouldOpenPopup = false;
+            ImGui::OpenPopup("##hex_editor_popup");
+        }
+
+        ImGui::SetNextWindowPos(ImGui::GetWindowPos() + ImGui::GetWindowContentRegionMin() - ImGui::GetStyle().WindowPadding);
+        if (ImGui::BeginPopup("##hex_editor_popup", ImGuiWindowFlags_NoTitleBar)) {
+
+            if (this->m_currPopup != nullptr)
+                this->m_currPopup->draw(this);
+            else
+                ImGui::CloseCurrentPopup();
+
+            ImGui::EndPopup();
+        } else {
+            this->closePopup();
+        }
+
+        // Right click menu
+        if (ImGui::IsMouseReleased(ImGuiMouseButton_Right) && ImGui::IsWindowHovered(ImGuiHoveredFlags_ChildWindows))
+            ImGui::OpenPopup("hex.builtin.menu.edit"_lang);
+
+        if (ImGui::BeginPopup("hex.builtin.menu.edit"_lang)) {
+            this->drawEditMenu();
+            ImGui::EndPopup();
+        }
     }
 
     void ViewHexEditor::drawContent() {
-        auto provider = ImHexApi::Provider::get();
+        if (ImGui::Begin(View::toWindowName(this->getUnlocalizedName()).c_str(), &this->getWindowOpenState(), ImGuiWindowFlags_NoCollapse | ImGuiWindowFlags_NoNavInputs | ImGuiWindowFlags_NoScrollbar | ImGuiWindowFlags_NoScrollWithMouse)) {
 
-        size_t dataSize = (!ImHexApi::Provider::isValid() || !provider->isReadable()) ? 0x00 : provider->getSize();
+            const float SeparatorColumWidth   = 6_scaled;
+            const auto CharacterSize          = ImGui::CalcTextSize("0");
 
-        this->m_memoryEditor.DrawWindow(View::toWindowName("hex.builtin.view.hex_editor.name").c_str(), &this->getWindowOpenState(), this, dataSize, dataSize == 0 ? 0x00 : provider->getBaseAddress() + provider->getCurrentPageAddress());
+            const color_t SelectionColor = 0x60C08080;
 
-        if (dataSize != 0x00) {
-            this->m_memoryEditor.OptShowAdvancedDecoding = this->m_advancedDecodingEnabled && this->m_currEncodingFile.valid();
+            const auto FooterSize = ImVec2(ImGui::GetContentRegionAvailWidth(), ImGui::GetTextLineHeightWithSpacing() * 3);
+            const auto TableSize = ImGui::GetContentRegionAvail() - ImVec2(0, FooterSize.y);
 
-            if (ImGui::Begin(View::toWindowName("hex.builtin.view.hex_editor.name").c_str())) {
+            const auto bytesPerCell = this->m_currDataVisualizer->getBytesPerCell();
+            const u16 columnCount = this->m_bytesPerRow / bytesPerCell;
+            const auto byteColumnCount = columnCount + getByteColumnSeparatorCount(columnCount);
 
-                if (ImGui::IsMouseReleased(ImGuiMouseButton_Right) && ImGui::IsWindowHovered(ImGuiHoveredFlags_ChildWindows))
-                    ImGui::OpenPopup("hex.builtin.menu.edit"_lang);
+            const auto selectionMin = std::min(this->m_selectionStart, this->m_selectionEnd);
+            const auto selectionMax = std::max(this->m_selectionStart, this->m_selectionEnd);
 
-                if (ImGui::BeginPopup("hex.builtin.menu.edit"_lang)) {
-                    this->drawEditPopup();
-                    ImGui::EndPopup();
+            this->drawPopup();
+
+            ImGui::PushStyleVar(ImGuiStyleVar_CellPadding, ImVec2(0.5, 0));
+            if (ImGui::BeginTable("##hex", 2 + byteColumnCount + 2 + 2 , ImGuiTableFlags_ScrollY | ImGuiTableFlags_SizingFixedFit | ImGuiTableFlags_NoKeepColumnsVisible, TableSize)) {
+                View::discardNavigationRequests();
+                ImGui::TableSetupScrollFreeze(0, 2);
+
+                // Row address column
+                ImGui::TableSetupColumn("Address");
+                ImGui::TableSetupColumn("");
+
+                // Byte columns
+                for (u16 i = 0; i < columnCount; i++) {
+                    if (isColumnSeparatorColumn(i, columnCount))
+                        ImGui::TableSetupColumn("", ImGuiTableColumnFlags_WidthFixed, SeparatorColumWidth);
+
+                    ImGui::TableSetupColumn(hex::format(this->m_upperCaseHex ? "{:0{}X}" : "{:0{}x}", i * bytesPerCell, this->m_currDataVisualizer->getMaxCharsPerCell()).c_str(), ImGuiTableColumnFlags_WidthFixed, CharacterSize.x * this->m_currDataVisualizer->getMaxCharsPerCell() + 6);
                 }
 
-                if (provider->getPageCount() > 1) {
-                    ImGui::NewLine();
+                // ASCII column
+                ImGui::TableSetupColumn("");
+                ImGui::TableSetupColumn("", ImGuiTableColumnFlags_WidthFixed, CharacterSize.x * this->m_bytesPerRow);
 
-                    auto linePos = ImGui::GetCursorPosY() - 15_scaled;
+                // Custom encoding column
+                ImGui::TableSetupColumn("");
+                ImGui::TableSetupColumn("", ImGuiTableColumnFlags_WidthStretch);
 
-                    ImGui::SetCursorPosY(linePos);
-
-                    if (ImGui::ArrowButton("prevPage", ImGuiDir_Left)) {
-                        provider->setCurrentPage(provider->getCurrentPage() - 1);
-
-                        EventManager::post<EventRegionSelected>(Region { std::min(this->m_memoryEditor.DataPreviewAddr, this->m_memoryEditor.DataPreviewAddrEnd), 1 });
-                    }
-
-                    ImGui::SameLine();
-
-                    if (ImGui::ArrowButton("nextPage", ImGuiDir_Right)) {
-                        provider->setCurrentPage(provider->getCurrentPage() + 1);
-
-                        EventManager::post<EventRegionSelected>(Region { std::min(this->m_memoryEditor.DataPreviewAddr, this->m_memoryEditor.DataPreviewAddrEnd), 1 });
-                    }
-
-                    ImGui::SameLine();
-                    ImGui::SeparatorEx(ImGuiSeparatorFlags_Vertical);
-                    ImGui::SameLine();
-                    ImGui::SetCursorPosY(linePos);
-
-                    ImGui::TextFormatted("hex.builtin.view.hex_editor.page"_lang, provider->getCurrentPage() + 1, provider->getPageCount());
+                ImGui::TableNextRow();
+                for (i32 i = 0; i < ImGui::TableGetColumnCount(); i++) {
+                    ImGui::TableNextColumn();
+                    ImGui::TextUnformatted(ImGui::TableGetColumnName(i));
+                    ImGui::SetCursorPosY(ImGui::GetCursorPosY() + CharacterSize.y / 2);
                 }
 
-                this->drawSearchPopup();
-                this->drawGotoPopup();
+                ImGui::TableNextRow();
+                ImGui::TableNextColumn();
+
+                if (ImHexApi::Provider::isValid()) {
+                    auto provider = ImHexApi::Provider::get();
+
+                    ImGuiListClipper clipper(std::ceil(provider->getSize() / (long double)(this->m_bytesPerRow)), CharacterSize.y);
+
+                    while (clipper.Step()) {
+                        this->m_visibleRowCount = clipper.DisplayEnd - clipper.DisplayStart;
+
+                        // Loop over rows
+                        for (i128 y = clipper.DisplayStart; y < u64(clipper.DisplayEnd); y++) {
+
+                            // Draw address column
+                            ImGui::TableNextRow();
+                            ImGui::TableNextColumn();
+                            ImGui::TextFormatted(this->m_upperCaseHex ? "{:08X}: " : "{:08x}: ", y * this->m_bytesPerRow + provider->getBaseAddress() + provider->getCurrentPageAddress());
+                            ImGui::TableNextColumn();
+
+                            const u8 validBytes = std::min<u64>(this->m_bytesPerRow, provider->getSize() - y * this->m_bytesPerRow);
+
+                            std::vector<u8> bytes(validBytes);
+                            provider->read(y * this->m_bytesPerRow + provider->getCurrentPageAddress(), bytes.data(), validBytes);
+
+                            std::vector<std::tuple<std::optional<color_t>, std::optional<color_t>>> cellColors;
+                            {
+                                for (u64 x = 0; x < columnCount; x++) {
+                                    const u64 byteAddress = y * this->m_bytesPerRow + x * bytesPerCell + provider->getBaseAddress() + provider->getCurrentPageAddress();
+
+                                    const auto cellBytes = std::min<u64>(validBytes, bytesPerCell);
+
+                                    // Query cell colors
+                                    const auto foregroundColor = queryForegroundColor(byteAddress, &bytes[x], cellBytes);
+                                    const auto backgroundColor = [&]{
+                                        auto color = queryBackgroundColor(byteAddress, &bytes[x], cellBytes);
+
+                                        if ((byteAddress >= selectionMin && byteAddress <= selectionMax)) {
+                                            if (color.has_value())
+                                                color = ImAlphaBlendColors(color.value(), SelectionColor);
+                                            else
+                                                color = SelectionColor;
+
+                                            *color = (*color & 0x00FFFFFF) | (this->m_highlightAlpha << 24);
+                                        }
+
+                                        return color;
+                                    }();
+
+                                    cellColors.emplace_back(
+                                        foregroundColor,
+                                        backgroundColor
+                                    );
+                                }
+                            }
+
+                            // Draw byte columns
+                            ImGui::PushStyleVar(ImGuiStyleVar_CellPadding, ImVec2(3, 0));
+
+                            bool shouldScroll = false;
+                            for (u64 x = 0; x < columnCount; x++) {
+                                const u64 byteAddress = y * this->m_bytesPerRow + x * bytesPerCell + provider->getBaseAddress() + provider->getCurrentPageAddress();
+
+                                ImGui::TableNextColumn();
+                                if (isColumnSeparatorColumn(x, columnCount))
+                                    ImGui::TableNextColumn();
+
+                                if (x < validBytes) {
+                                    auto cellStartPos  = (ImGui::GetWindowPos() + ImGui::GetCursorPos()) - ImGui::GetStyle().CellPadding - ImVec2(ImGui::GetScrollX(), ImGui::GetScrollY());
+                                    auto cellSize = (CharacterSize * ImVec2(this->m_currDataVisualizer->getMaxCharsPerCell(), 1) + (ImVec2(3, 2) * ImGui::GetStyle().CellPadding) - ImVec2(1, 0) * ImGui::GetStyle().CellPadding) + ImVec2(1, 0);
+
+                                    auto [foregroundColor, backgroundColor] = cellColors[x];
+
+                                    if (isColumnSeparatorColumn(x + 1, columnCount) && selectionMax != x + y * columnCount) {
+                                        cellSize.x += SeparatorColumWidth + 1;
+                                    }
+                                    if (y == clipper.DisplayStart)
+                                        cellSize.y -= (ImGui::GetStyle().CellPadding.y + 1);
+
+                                    // Draw highlights and selection
+                                    if (backgroundColor.has_value()) {
+                                        auto drawList = ImGui::GetWindowDrawList();
+
+                                        // Draw background color
+                                        drawList->AddRectFilled(cellStartPos, cellStartPos + cellSize, backgroundColor.value());
+
+                                        // Draw frame around mouse selection
+                                        this->drawSelectionFrame(x, y, byteAddress, bytesPerCell, cellStartPos, cellSize);
+                                    }
+
+
+                                    const bool cellHovered = ImGui::IsMouseHoveringRect(cellStartPos, cellStartPos + cellSize, false);
+
+                                    // Handle selection
+                                    {
+                                        if (ImGui::IsWindowHovered() && cellHovered && ImGui::IsMouseHoveringRect(ImGui::GetWindowPos(), ImGui::GetWindowPos() + TableSize)) {
+                                            drawTooltip(byteAddress, &bytes[x], bytesPerCell);
+
+                                            auto endAddress = byteAddress + bytesPerCell - 1;
+                                            if (ImGui::IsMouseDragging(ImGuiMouseButton_Left)) {
+                                                this->setSelection(this->m_selectionStart, endAddress);
+                                                shouldScroll = true;
+                                            }
+                                            else if (ImGui::IsMouseDown(ImGuiMouseButton_Left)) {
+                                                if (ImGui::GetIO().KeyShift)
+                                                    this->setSelection(this->m_selectionStart, endAddress);
+                                                else
+                                                    this->setSelection(byteAddress, endAddress);
+
+                                                shouldScroll = true;
+                                            }
+                                        }
+
+                                    }
+
+                                    // Get byte foreground color
+                                    if (foregroundColor.has_value())
+                                        ImGui::PushStyleColor(ImGuiCol_Text, *foregroundColor);
+
+                                    // Draw cell content
+                                    ImGui::PushStyleVar(ImGuiStyleVar_FramePadding, ImVec2(0, 0));
+                                    ImGui::PushItemWidth((CharacterSize * this->m_currDataVisualizer->getMaxCharsPerCell()).x);
+                                    this->drawCell(byteAddress, &bytes[x], bytesPerCell, cellHovered);
+                                    ImGui::PopItemWidth();
+                                    ImGui::PopStyleVar();
+
+                                    if (foregroundColor.has_value())
+                                        ImGui::PopStyleColor();
+                                }
+                            }
+                            ImGui::PopStyleVar();
+
+                            // Scroll to the cursor if it's either at the top or bottom edge of the screen
+                            if (shouldScroll && this->m_selectionEnd != InvalidSelection && this->m_selectionStart != this->m_selectionEnd) {
+                                if (y == clipper.DisplayStart) {
+                                    if (i128(this->m_selectionEnd - provider->getBaseAddress() - provider->getCurrentPageAddress()) <= (i64(clipper.DisplayStart + 4) * this->m_bytesPerRow))
+                                        ImGui::SetScrollHereY(0.1);
+                                } else if (y == (clipper.DisplayEnd - 1)) {
+                                    if (i128(this->m_selectionEnd - provider->getBaseAddress() - provider->getCurrentPageAddress()) >= (i64(clipper.DisplayEnd - 2) * this->m_bytesPerRow))
+                                        ImGui::SetScrollHereY(0.95);
+                                }
+                            }
+
+                            ImGui::TableNextColumn();
+                            ImGui::TableNextColumn();
+
+                            // Draw ASCII column
+                            if (this->m_showAscii) {
+                                ImGui::PushStyleVar(ImGuiStyleVar_CellPadding, ImVec2(0, 0));
+                                if (ImGui::BeginTable("##ascii_cell", this->m_bytesPerRow)) {
+                                    ImGui::TableNextRow();
+
+                                    for (u64 x = 0; x < this->m_bytesPerRow; x++) {
+                                        ImGui::TableNextColumn();
+
+                                        const u64 byteAddress = y * this->m_bytesPerRow + x + provider->getBaseAddress() + provider->getCurrentPageAddress();
+
+                                        const auto cellStartPos  = (ImGui::GetWindowPos() + ImGui::GetCursorPos()) - ImGui::GetStyle().CellPadding - ImVec2(ImGui::GetScrollX(), ImGui::GetScrollY());
+                                        const auto cellSize = CharacterSize;
+
+                                        const bool cellHovered = ImGui::IsMouseHoveringRect(cellStartPos, cellStartPos + cellSize, false);
+
+                                        // Handle selection
+                                        {
+                                            if (ImGui::IsWindowHovered() && cellHovered && ImGui::IsMouseHoveringRect(ImGui::GetWindowPos(), ImGui::GetWindowPos() + TableSize)) {
+                                                drawTooltip(byteAddress, &bytes[x], bytesPerCell);
+                                                if (ImGui::IsMouseDragging(ImGuiMouseButton_Left)) {
+                                                    this->setSelection(this->m_selectionStart, byteAddress);
+                                                }
+                                                else if (ImGui::IsMouseDown(ImGuiMouseButton_Left)) {
+                                                    if (ImGui::GetIO().KeyShift)
+                                                        this->setSelection(this->m_selectionStart, byteAddress);
+                                                    else
+                                                        this->setSelection(byteAddress, byteAddress);
+                                                }
+                                            }
+                                        }
+
+                                        auto [foregroundColor, backgroundColor] = cellColors[x / bytesPerCell];
+
+                                        // Draw highlights and selection
+                                        if (backgroundColor.has_value()) {
+                                            auto drawList = ImGui::GetWindowDrawList();
+
+                                            // Draw background color
+                                            drawList->AddRectFilled(cellStartPos, cellStartPos + cellSize, backgroundColor.value());
+
+                                            this->drawSelectionFrame(x, y, byteAddress, 1, cellStartPos, cellSize);
+                                        }
+
+                                        if (x < validBytes) {
+                                            if (std::isprint(bytes[x]))
+                                                ImGui::TextFormatted("{:c}", bytes[x]);
+                                            else
+                                                ImGui::TextDisabled(".");
+                                        }
+                                    }
+
+                                    ImGui::EndTable();
+                                }
+                                ImGui::PopStyleVar();
+                            }
+
+                            ImGui::TableNextColumn();
+                            ImGui::TableNextColumn();
+
+                            // Draw Custom encoding column
+                            if (this->m_currCustomEncoding.has_value()) {
+                                std::vector<std::pair<u64, CustomEncodingData>> encodingData;
+                                u32 offset = 0;
+                                do {
+                                    const u64 address = y * this->m_bytesPerRow + offset + provider->getBaseAddress() + provider->getCurrentPageAddress();
+
+                                    auto result = queryCustomEncodingData(*this->m_currCustomEncoding, address);
+                                    offset += std::max<size_t>(1, result.advance);
+
+                                    encodingData.emplace_back(address, result);
+                                } while (offset < this->m_bytesPerRow);
+
+                                ImGui::PushStyleVar(ImGuiStyleVar_CellPadding, ImVec2(0, 0));
+                                if (ImGui::BeginTable("##encoding_cell", encodingData.size(), ImGuiTableFlags_SizingFixedFit)) {
+                                    ImGui::TableNextRow();
+
+                                    for (const auto &[address, data] : encodingData) {
+                                        ImGui::TableNextColumn();
+
+                                        const auto cellStartPos  = (ImGui::GetWindowPos() + ImGui::GetCursorPos()) - ImGui::GetStyle().CellPadding - ImVec2(ImGui::GetScrollX(), ImGui::GetScrollY());
+                                        const auto cellSize = CharacterSize * data.advance;
+                                        const bool cellHovered = ImGui::IsMouseHoveringRect(cellStartPos, cellStartPos + cellSize, false);
+
+                                        auto [foregroundColor, backgroundColor] = cellColors[address % this->m_bytesPerRow];
+
+                                        // Draw highlights and selection
+                                        if (backgroundColor.has_value()) {
+                                            auto drawList = ImGui::GetWindowDrawList();
+
+                                            // Draw background color
+                                            drawList->AddRectFilled(cellStartPos, cellStartPos + cellSize, backgroundColor.value());
+
+                                            this->drawSelectionFrame(address % this->m_bytesPerRow, y, address, 1, cellStartPos, cellSize);
+                                        }
+
+                                        ImGui::TextFormattedColored(data.color, "{}", data.displayValue);
+
+                                        // Handle selection
+                                        {
+                                            if (ImGui::IsWindowHovered() && cellHovered && ImGui::IsMouseHoveringRect(ImGui::GetWindowPos(), ImGui::GetWindowPos() + TableSize)) {
+                                                auto endAddress = address + data.advance - 1;
+
+                                                if (ImGui::IsMouseDragging(ImGuiMouseButton_Left)) {
+                                                    this->setSelection(this->m_selectionStart, endAddress);
+                                                }
+                                                else if (ImGui::IsMouseDown(ImGuiMouseButton_Left)) {
+                                                    if (ImGui::GetIO().KeyShift)
+                                                        this->setSelection(this->m_selectionStart, endAddress);
+                                                    else
+                                                        this->setSelection(address, endAddress);
+                                                }
+                                            }
+                                        }
+                                    }
+
+                                    ImGui::EndTable();
+                                }
+                                ImGui::PopStyleVar();
+
+                            }
+                        }
+                    }
+                } else {
+                    ImGui::TextFormattedCentered("No bytes available");
+                }
+
+                ImGui::EndTable();
             }
-            ImGui::End();
+            ImGui::PopStyleVar();
+
+            if (ImHexApi::Provider::isValid()) {
+                auto provider = ImHexApi::Provider::get();
+                const auto pageCount = provider->getPageCount();
+                constexpr static u32 MinPage = 1;
+
+                const auto windowEndPos = ImGui::GetWindowPos() + ImGui::GetWindowSize() - ImGui::GetStyle().WindowPadding;
+                ImGui::GetWindowDrawList()->AddLine(windowEndPos - ImVec2(0, FooterSize.y - 1_scaled), windowEndPos - FooterSize + ImVec2(0, 1_scaled), ImGui::GetColorU32(ImGuiCol_Separator), 2.0_scaled);
+
+                if (ImGui::BeginChild("##footer", FooterSize, false, ImGuiWindowFlags_NoScrollbar | ImGuiWindowFlags_NoScrollWithMouse)) {
+                    if (ImGui::BeginTable("##footer_table", 2)) {
+                        ImGui::TableNextRow();
+
+                        // Page slider
+                        ImGui::TableNextColumn();
+                        {
+                            u32 page = provider->getCurrentPage() + 1;
+                            ImGui::BeginDisabled(pageCount <= 1);
+                            if (ImGui::SliderScalar("Page", ImGuiDataType_U32, &page, &MinPage, &pageCount, hex::format("%d / {}", pageCount).c_str()))
+                                provider->setCurrentPage(page - 1);
+
+                            ImGui::EndDisabled();
+                        }
+
+                        // Page Address
+                        ImGui::TableNextColumn();
+                        {
+                            ImGui::TextFormatted("Region: 0x{0:08X} - 0x{1:08X} ({0} - {1})", provider->getCurrentPageAddress(), provider->getSize());
+                        }
+
+                        ImGui::TableNextRow();
+
+                        // Selection
+                        ImGui::TableNextColumn();
+                        {
+                            if (selectionMin == InvalidSelection || selectionMax == InvalidSelection)
+                                ImGui::TextUnformatted("Selection: None");
+                            else
+                                ImGui::TextFormatted("Selection: 0x{0:08X} - 0x{1:08X} ({0} - {1})", selectionMin, selectionMax);
+                        }
+
+                        // Loaded data size
+                        ImGui::TableNextColumn();
+                        {
+                            ImGui::TextFormatted("Data size: 0x{0:08X} ({1})", provider->getActualSize(), hex::toByteString(provider->getActualSize()));
+                        }
+
+                        ImGui::EndTable();
+                    }
+
+                    ImGui::EndChild();
+                }
+            }
+
+
+            // Handle jumping to selection
+            if (this->m_shouldScrollToSelection) {
+                this->m_shouldScrollToSelection = false;
+                ImGui::BeginChild("##hex");
+                ImGui::SetScrollFromPosY(ImGui::GetCursorStartPos().y + (float(this->m_selectionStart) / this->m_bytesPerRow) * CharacterSize.y, 0.0);
+                ImGui::EndChild();
+            }
+
         }
+        ImGui::End();
+
+        this->m_selectionChanged = false;
+    }
+
+    void ViewHexEditor::drawAlwaysVisible() {
+
     }
 
     static void save() {
@@ -274,154 +948,11 @@ namespace hex::plugin::builtin {
         });
     }
 
-    void ViewHexEditor::drawAlwaysVisible() {
+    static void copyBytes(const Region &selection) {
         auto provider = ImHexApi::Provider::get();
 
-        if (ImGui::BeginPopupModal("hex.builtin.view.hex_editor.exit_application.title"_lang, nullptr, ImGuiWindowFlags_AlwaysAutoResize)) {
-            ImGui::NewLine();
-            ImGui::TextUnformatted("hex.builtin.view.hex_editor.exit_application.desc"_lang);
-            ImGui::NewLine();
-
-            confirmButtons(
-                "hex.builtin.common.yes"_lang, "hex.builtin.common.no"_lang, [] { ImHexApi::Common::closeImHex(true); }, [] { ImGui::CloseCurrentPopup(); });
-
-            if (ImGui::IsKeyDown(ImGui::GetKeyIndex(ImGuiKey_Escape)))
-                ImGui::CloseCurrentPopup();
-
-            ImGui::EndPopup();
-        }
-
-        if (ImGui::BeginPopupModal("hex.builtin.view.hex_editor.script.title"_lang, nullptr, ImGuiWindowFlags_AlwaysAutoResize)) {
-            ImGui::SetCursorPosX(10);
-            ImGui::TextFormattedWrapped("{}", static_cast<const char *>("hex.builtin.view.hex_editor.script.desc"_lang));
-
-            ImGui::NewLine();
-            ImGui::InputText("##nolabel", this->m_loaderScriptScriptPath.data(), this->m_loaderScriptScriptPath.length(), ImGuiInputTextFlags_ReadOnly);
-            ImGui::SameLine();
-            if (ImGui::Button("hex.builtin.view.hex_editor.script.script"_lang)) {
-                fs::openFileBrowser(fs::DialogMode::Open, { {"Python Script", "py"} },
-                    [this](const auto &path) {
-                        this->m_loaderScriptScriptPath = path.string();
-                    });
-            }
-            ImGui::InputText("##nolabel", this->m_loaderScriptFilePath.data(), this->m_loaderScriptFilePath.length(), ImGuiInputTextFlags_ReadOnly);
-            ImGui::SameLine();
-            if (ImGui::Button("hex.builtin.view.hex_editor.script.file"_lang)) {
-                fs::openFileBrowser(fs::DialogMode::Open, {}, [this](const auto &path) {
-                    this->m_loaderScriptFilePath = path.string();
-                });
-            }
-            if (ImGui::IsKeyDown(ImGui::GetKeyIndex(ImGuiKey_Escape)))
-                ImGui::CloseCurrentPopup();
-
-            ImGui::NewLine();
-
-            confirmButtons(
-                "hex.builtin.common.load"_lang, "hex.builtin.common.cancel"_lang, [this, &provider] {
-                   if (!this->m_loaderScriptScriptPath.empty() && !this->m_loaderScriptFilePath.empty()) {
-                       EventManager::post<RequestOpenFile>(this->m_loaderScriptFilePath);
-                       LoaderScript::setFilePath(this->m_loaderScriptFilePath);
-                       LoaderScript::setDataProvider(provider);
-                       LoaderScript::processFile(this->m_loaderScriptScriptPath);
-                       ImGui::CloseCurrentPopup();
-                   } }, [] { ImGui::CloseCurrentPopup(); });
-
-            ImGui::EndPopup();
-        }
-
-        if (ImGui::BeginPopupModal("hex.builtin.view.hex_editor.menu.edit.set_base"_lang, nullptr, ImGuiWindowFlags_AlwaysAutoResize)) {
-            ImGui::InputHexadecimal("hex.builtin.common.address"_lang, &this->m_baseAddress, ImGuiInputTextFlags_CharsHexadecimal);
-            ImGui::NewLine();
-
-            confirmButtons(
-                "hex.builtin.common.set"_lang, "hex.builtin.common.cancel"_lang, [this, &provider] {
-                               provider->setBaseAddress(this->m_baseAddress);
-                               ImGui::CloseCurrentPopup(); }, [] { ImGui::CloseCurrentPopup(); });
-
-            if (ImGui::IsKeyDown(ImGui::GetKeyIndex(ImGuiKey_Escape)))
-                ImGui::CloseCurrentPopup();
-
-            ImGui::EndPopup();
-        }
-
-        if (ImGui::BeginPopupModal("hex.builtin.view.hex_editor.menu.edit.resize"_lang, nullptr, ImGuiWindowFlags_AlwaysAutoResize)) {
-            ImGui::InputHexadecimal("hex.builtin.common.size"_lang, &this->m_resizeSize, ImGuiInputTextFlags_CharsHexadecimal);
-            ImGui::NewLine();
-
-            confirmButtons(
-                "hex.builtin.common.set"_lang, "hex.builtin.common.cancel"_lang, [this, &provider] {
-                    provider->resize(this->m_resizeSize);
-                    ImGui::CloseCurrentPopup(); }, [] { ImGui::CloseCurrentPopup(); });
-
-            if (ImGui::IsKeyDown(ImGui::GetKeyIndex(ImGuiKey_Escape)))
-                ImGui::CloseCurrentPopup();
-
-            ImGui::EndPopup();
-        }
-
-        if (ImGui::BeginPopupModal("hex.builtin.view.hex_editor.menu.edit.insert"_lang, nullptr, ImGuiWindowFlags_AlwaysAutoResize)) {
-            ImGui::InputHexadecimal("hex.builtin.common.size"_lang, &this->m_resizeSize, ImGuiInputTextFlags_CharsHexadecimal);
-            ImGui::NewLine();
-
-            confirmButtons(
-                "hex.builtin.common.set"_lang, "hex.builtin.common.cancel"_lang, [this, &provider] {
-                    provider->insert(std::min(this->m_memoryEditor.DataPreviewAddr, this->m_memoryEditor.DataPreviewAddrEnd), this->m_resizeSize);
-                    ImGui::CloseCurrentPopup(); }, [] { ImGui::CloseCurrentPopup(); });
-
-            if (ImGui::IsKeyDown(ImGui::GetKeyIndex(ImGuiKey_Escape)))
-                ImGui::CloseCurrentPopup();
-
-            ImGui::EndPopup();
-        }
-    }
-
-    void ViewHexEditor::openFile(const std::fs::path &path) {
-        hex::prv::Provider *provider = nullptr;
-        EventManager::post<RequestCreateProvider>("hex.builtin.provider.file", &provider);
-
-        if (auto fileProvider = dynamic_cast<prv::FileProvider *>(provider)) {
-            fileProvider->setPath(path);
-            if (!fileProvider->open()) {
-                View::showErrorPopup("hex.builtin.view.hex_editor.error.open"_lang);
-                ImHexApi::Provider::remove(provider);
-
-                return;
-            }
-        }
-
-        if (!provider->isWritable()) {
-            this->m_memoryEditor.ReadOnly = true;
-            View::showErrorPopup("hex.builtin.view.hex_editor.error.read_only"_lang);
-        } else {
-            this->m_memoryEditor.ReadOnly = false;
-        }
-
-        if (!provider->isAvailable()) {
-            View::showErrorPopup("hex.builtin.view.hex_editor.error.open"_lang);
-            ImHexApi::Provider::remove(provider);
-
-            return;
-        }
-
-        ProjectFile::setFilePath(path);
-
-        this->getWindowOpenState() = true;
-
-        EventManager::post<EventFileLoaded>(path);
-        EventManager::post<EventDataChanged>();
-        EventManager::post<EventHighlightingChanged>();
-    }
-
-    void ViewHexEditor::copyBytes() const {
-        auto provider = ImHexApi::Provider::get();
-
-        size_t start = std::min(this->m_memoryEditor.DataPreviewAddr, this->m_memoryEditor.DataPreviewAddrEnd);
-        size_t end   = std::max(this->m_memoryEditor.DataPreviewAddr, this->m_memoryEditor.DataPreviewAddrEnd);
-
-        size_t copySize = (end - start) + 1;
-
-        std::vector<u8> buffer(copySize, 0x00);
-        provider->read(start + provider->getBaseAddress() + provider->getCurrentPageAddress(), buffer.data(), buffer.size());
+        std::vector<u8> buffer(selection.size, 0x00);
+        provider->read(selection.getStartAddress() + provider->getBaseAddress() + provider->getCurrentPageAddress(), buffer.data(), buffer.size());
 
         std::string str;
         for (const auto &byte : buffer)
@@ -431,11 +962,8 @@ namespace hex::plugin::builtin {
         ImGui::SetClipboardText(str.c_str());
     }
 
-    void ViewHexEditor::pasteBytes() const {
+    static void pasteBytes(const Region &selection) {
         auto provider = ImHexApi::Provider::get();
-
-        size_t start = std::min(this->m_memoryEditor.DataPreviewAddr, this->m_memoryEditor.DataPreviewAddrEnd);
-        size_t end   = std::max(this->m_memoryEditor.DataPreviewAddr, this->m_memoryEditor.DataPreviewAddrEnd);
 
         std::string clipboard = ImGui::GetClipboardText();
 
@@ -470,461 +998,154 @@ namespace hex::plugin::builtin {
         }
 
         // Write bytes
-        provider->write(start + provider->getBaseAddress() + provider->getCurrentPageAddress(), buffer.data(), std::min(end - start + 1, buffer.size()));
+        provider->write(selection.getStartAddress() + provider->getBaseAddress() + provider->getCurrentPageAddress(), buffer.data(), std::min(selection.size, buffer.size()));
     }
 
-    void ViewHexEditor::copyString() const {
+    static void copyString(const Region &selection) {
         auto provider = ImHexApi::Provider::get();
 
-        size_t start = std::min(this->m_memoryEditor.DataPreviewAddr, this->m_memoryEditor.DataPreviewAddrEnd);
-        size_t end   = std::max(this->m_memoryEditor.DataPreviewAddr, this->m_memoryEditor.DataPreviewAddrEnd);
-
-        size_t copySize = (end - start) + 1;
-
-        std::string buffer(copySize, 0x00);
-        buffer.reserve(copySize);
-        provider->read(start + provider->getBaseAddress() + provider->getCurrentPageAddress(), buffer.data(), copySize);
+        std::string buffer(selection.size, 0x00);
+        buffer.reserve(selection.size);
+        provider->read(selection.getStartAddress() + provider->getBaseAddress() + provider->getCurrentPageAddress(), buffer.data(), selection.size);
 
         ImGui::SetClipboardText(buffer.c_str());
     }
 
-    static std::vector<std::pair<u64, u64>> findString(hex::prv::Provider *&provider, std::string string) {
-        std::vector<std::pair<u64, u64>> results;
+    void ViewHexEditor::registerShortcuts() {
 
-        if (string.empty())
-            return {};
+        ShortcutManager::addShortcut(this, CTRL + Keys::S, [] {
+            save();
+        });
+        ShortcutManager::addShortcut(this, CTRL + SHIFT + Keys::S, [] {
+            saveAs();
+        });
 
-        u32 foundCharacters = 0;
+        ShortcutManager::addShortcut(this, CTRL + Keys::A, [this] {
+            if (ImHexApi::Provider::isValid())
+                this->setSelection(size_t(0), ImHexApi::Provider::get()->getActualSize());
+        });
 
-        std::vector<u8> buffer(1024, 0x00);
-        size_t dataSize = provider->getSize();
-        for (u64 offset = 0; offset < dataSize; offset += 1024) {
-            size_t usedBufferSize = std::min<size_t>(buffer.size(), dataSize - offset);
-            provider->read(offset + provider->getBaseAddress() + provider->getCurrentPageAddress(), buffer.data(), usedBufferSize);
+        ShortcutManager::addShortcut(this, Keys::Escape, [this] {
+            this->m_selectionStart = InvalidSelection;
+            this->m_selectionEnd = InvalidSelection;
+            EventManager::post<EventRegionSelected>(this->getSelection());
+        });
 
-            for (u64 i = 0; i < usedBufferSize; i++) {
-                if (buffer[i] == string[foundCharacters])
-                    foundCharacters++;
-                else
-                    foundCharacters = 0;
-
-                if (foundCharacters == string.size()) {
-                    results.emplace_back(offset + i - foundCharacters + 1, offset + i);
-                    foundCharacters = 0;
-                }
+        ShortcutManager::addShortcut(this, Keys::Up, [this] {
+            if (this->m_selectionStart >= this->m_bytesPerRow) {
+                auto pos = this->m_selectionStart - this->m_bytesPerRow;
+                this->setSelection(pos, pos);
             }
-        }
-
-        return results;
-    }
-
-    static std::vector<std::pair<u64, u64>> findHex(hex::prv::Provider *&provider, std::string string) {
-        std::vector<std::pair<u64, u64>> results;
-
-        if ((string.size() % 2) == 1)
-            string = "0" + string;
-
-        if (string.empty())
-            return {};
-
-        std::vector<u8> hex;
-        hex.reserve(string.size() / 2);
-
-        for (u32 i = 0; i < string.size(); i += 2) {
-            char byte[3] = { string[i], string[i + 1], 0 };
-            hex.push_back(strtoul(byte, nullptr, 16));
-        }
-
-        u32 foundCharacters = 0;
-
-        std::vector<u8> buffer(1024, 0x00);
-        size_t dataSize = provider->getSize();
-        for (u64 offset = 0; offset < dataSize; offset += 1024) {
-            size_t usedBufferSize = std::min<size_t>(buffer.size(), dataSize - offset);
-            provider->read(offset + provider->getBaseAddress() + provider->getCurrentPageAddress(), buffer.data(), usedBufferSize);
-
-            for (u64 i = 0; i < usedBufferSize; i++) {
-                if (buffer[i] == hex[foundCharacters])
-                    foundCharacters++;
-                else
-                    foundCharacters = 0;
-
-                if (foundCharacters == hex.size()) {
-                    results.emplace_back(offset + i - foundCharacters + 1, offset + i);
-                    foundCharacters = 0;
-                }
+        });
+        ShortcutManager::addShortcut(this, Keys::Down, [this] {
+            auto pos = this->m_selectionStart + this->m_bytesPerRow;
+            this->setSelection(pos, pos);
+        });
+        ShortcutManager::addShortcut(this, Keys::Left, [this] {
+            if (this->m_selectionStart > 0) {
+                auto pos = this->m_selectionStart - 1;
+                this->setSelection(pos, pos);
             }
-        }
+        });
+        ShortcutManager::addShortcut(this, Keys::Right, [this] {
+            auto pos = this->m_selectionStart + 1;
+            this->setSelection(pos, pos);
+        });
 
-        return results;
-    }
-
-    void ViewHexEditor::drawSearchInput(std::vector<char> *currBuffer, ImGuiInputTextFlags flags) {
-        if (this->m_searchRequested) {
-            ImGui::SetKeyboardFocusHere();
-            this->m_searchRequested = false;
-        }
-
-        if (ImGui::InputText("##nolabel", currBuffer->data(), currBuffer->size(), flags, ViewHexEditor::inputCallback, this)) {
-            this->m_searchRequested = true;
-            if (this->m_lastSearchBuffer == nullptr || this->m_lastSearchBuffer->empty())
-                performSearch(currBuffer->data());
-            else
-                performSearchNext();
-        }
-
-        ImGui::EndTabItem();
-    }
-
-    void ViewHexEditor::performSearch(const char *buffer) {
-        auto provider = ImHexApi::Provider::get();
-
-        *this->m_lastSearchBuffer = this->m_searchFunction(provider, buffer);
-        this->m_lastSearchIndex   = 0;
-
-        if (!this->m_lastSearchBuffer->empty())
-            this->m_memoryEditor.GotoAddrAndSelect((*this->m_lastSearchBuffer)[0].first, (*this->m_lastSearchBuffer)[0].second);
-    }
-
-    void ViewHexEditor::performSearchNext() {
-        if (this->m_lastSearchBuffer->empty()) {
-            return;
-        }
-
-        ++this->m_lastSearchIndex %= this->m_lastSearchBuffer->size();
-        this->m_memoryEditor.GotoAddrAndSelect((*this->m_lastSearchBuffer)[this->m_lastSearchIndex].first,
-            (*this->m_lastSearchBuffer)[this->m_lastSearchIndex].second);
-    }
-
-    void ViewHexEditor::performSearchPrevious() {
-        if (this->m_lastSearchBuffer->empty()) {
-            return;
-        }
-
-        this->m_lastSearchIndex--;
-
-        if (this->m_lastSearchIndex < 0)
-            this->m_lastSearchIndex = this->m_lastSearchBuffer->size() - 1;
-
-        this->m_lastSearchIndex %= this->m_lastSearchBuffer->size();
-
-        this->m_memoryEditor.GotoAddrAndSelect((*this->m_lastSearchBuffer)[this->m_lastSearchIndex].first,
-            (*this->m_lastSearchBuffer)[this->m_lastSearchIndex].second);
-    }
-
-    int ViewHexEditor::inputCallback(ImGuiInputTextCallbackData *data) {
-        auto _this    = static_cast<ViewHexEditor *>(data->UserData);
-        auto provider = ImHexApi::Provider::get();
-
-        *_this->m_lastSearchBuffer = _this->m_searchFunction(provider, data->Buf);
-        _this->m_lastSearchIndex   = 0;
-
-        if (!_this->m_lastSearchBuffer->empty())
-            _this->m_memoryEditor.GotoAddrAndSelect((*_this->m_lastSearchBuffer)[0].first, (*_this->m_lastSearchBuffer)[0].second);
-
-        return 0;
-    }
-
-    void ViewHexEditor::drawSearchPopup() {
-        ImGui::SetNextWindowPos(ImGui::GetWindowPos() + ImGui::GetWindowContentRegionMin() - ImGui::GetStyle().WindowPadding);
-        if (ImGui::BeginPopup("hex.builtin.view.hex_editor.menu.file.search"_lang)) {
-            if (ImGui::BeginTabBar("searchTabs")) {
-                std::vector<char> *currBuffer = nullptr;
-                if (ImGui::BeginTabItem("hex.builtin.view.hex_editor.search.string"_lang)) {
-                    this->m_searchFunction   = findString;
-                    this->m_lastSearchBuffer = &this->m_lastStringSearch;
-                    currBuffer               = &this->m_searchStringBuffer;
-
-                    drawSearchInput(currBuffer, ImGuiInputTextFlags_CallbackCompletion | ImGuiInputTextFlags_EnterReturnsTrue);
-                }
-
-                if (ImGui::BeginTabItem("hex.builtin.view.hex_editor.search.hex"_lang)) {
-                    this->m_searchFunction   = findHex;
-                    this->m_lastSearchBuffer = &this->m_lastHexSearch;
-                    currBuffer               = &this->m_searchHexBuffer;
-
-                    drawSearchInput(currBuffer, ImGuiInputTextFlags_CharsHexadecimal | ImGuiInputTextFlags_CallbackCompletion | ImGuiInputTextFlags_EnterReturnsTrue);
-                }
-
-                if (currBuffer != nullptr) {
-                    if (ImGui::Button("hex.builtin.view.hex_editor.search.find"_lang))
-                        performSearch(currBuffer->data());
-
-                    if (!this->m_lastSearchBuffer->empty()) {
-                        if ((ImGui::Button("hex.builtin.view.hex_editor.search.find_next"_lang)))
-                            performSearchNext();
-
-                        ImGui::SameLine();
-
-                        if ((ImGui::Button("hex.builtin.view.hex_editor.search.find_prev"_lang)))
-                            performSearchPrevious();
-                    }
-                }
-
-                ImGui::EndTabBar();
+        ShortcutManager::addShortcut(this, Keys::PageUp, [this] {
+            u64 visibleByteCount = this->m_bytesPerRow * this->m_visibleRowCount;
+            if (this->m_selectionStart >= visibleByteCount) {
+                auto pos = this->m_selectionStart - visibleByteCount;
+                this->setSelection(pos, pos);
+                this->jumpToSelection();
             }
+        });
+        ShortcutManager::addShortcut(this, Keys::PageDown, [this] {
+            auto pos = this->m_selectionStart + (this->m_bytesPerRow * this->m_visibleRowCount);
+            this->setSelection(pos, pos);
+            this->jumpToSelection();
+        });
 
-            if (ImGui::IsKeyDown(ImGui::GetKeyIndex(ImGuiKey_Escape)))
-                ImGui::CloseCurrentPopup();
+        ShortcutManager::addShortcut(this, SHIFT + Keys::Up, [this] {
+            this->setSelection(this->m_selectionStart - this->m_bytesPerRow, this->m_selectionEnd);
+        });
+        ShortcutManager::addShortcut(this, SHIFT + Keys::Down, [this] {
+            this->setSelection(this->m_selectionStart + this->m_bytesPerRow, this->m_selectionEnd);
 
-            ImGui::EndPopup();
-        }
-    }
+        });
+        ShortcutManager::addShortcut(this, SHIFT + Keys::Left, [this] {
+            this->setSelection(this->m_selectionStart - 1, this->m_selectionEnd);
 
-    void ViewHexEditor::drawGotoPopup() {
-        auto provider    = ImHexApi::Provider::get();
-        auto baseAddress = provider->getBaseAddress();
-        auto dataSize    = provider->getActualSize();
+        });
+        ShortcutManager::addShortcut(this, SHIFT + Keys::Right, [this] {
+            this->setSelection(this->m_selectionStart + 1, this->m_selectionEnd);
+        });
 
-        ImGui::SetNextWindowPos(ImGui::GetWindowPos() + ImGui::GetWindowContentRegionMin() - ImGui::GetStyle().WindowPadding);
-        if (ImGui::BeginPopup("hex.builtin.view.hex_editor.menu.file.goto"_lang)) {
-            bool runGoto = this->m_evaluateGoto;
-            this->m_evaluateGoto = false;
+        ShortcutManager::addShortcut(this, CTRL + Keys::G, [this] {
+            if (!ImHexApi::Provider::isValid()) return;
 
-            if (ImGui::BeginTabBar("gotoTabs")) {
-                u64 newOffset = 0;
-                if (ImGui::BeginTabItem("hex.builtin.view.hex_editor.goto.offset.absolute"_lang)) {
-                    if (this->m_gotoRequested) {
-                        ImGui::SetKeyboardFocusHere();
-                        this->m_gotoRequested = false;
-                    }
+            this->openPopup<PopupGoto>();
+        });
+        ShortcutManager::addShortcut(this, CTRL + Keys::F, [this] {
+            if (!ImHexApi::Provider::isValid()) return;
 
-                    runGoto = ImGui::InputText("##goto", this->m_gotoAddressInput.data(), this->m_gotoAddressInput.size(), ImGuiInputTextFlags_EnterReturnsTrue | ImGuiInputTextFlags_CallbackResize, ImGui::UpdateStringSizeCallback, &this->m_gotoAddressInput) || runGoto;
-                    if (runGoto) {
-                        MathEvaluator evaluator;
-                        auto result = evaluator.evaluate(this->m_gotoAddressInput);
-                        if (result) {
-                            if (*result < baseAddress || *result > baseAddress + dataSize)
-                                newOffset = baseAddress;
-                            else
-                                newOffset = *result;
-                        } else {
-                            runGoto = false;
-                        }
-                    }
+            this->openPopup<PopupFind>();
+        });
 
-                    ImGui::EndTabItem();
-                }
-                if (ImGui::BeginTabItem("hex.builtin.view.hex_editor.goto.offset.begin"_lang)) {
-                    if (this->m_gotoRequested) {
-                        ImGui::SetKeyboardFocusHere();
-                        this->m_gotoRequested = false;
-                    }
-                    runGoto = ImGui::InputText("##goto", this->m_gotoAddressInput.data(), this->m_gotoAddressInput.size(), ImGuiInputTextFlags_EnterReturnsTrue | ImGuiInputTextFlags_CallbackResize, ImGui::UpdateStringSizeCallback, &this->m_gotoAddressInput) || runGoto;
-                    if (runGoto) {
-                        MathEvaluator evaluator;
-                        auto result = evaluator.evaluate(this->m_gotoAddressInput);
-                        if (result) {
-                            if (*result < 0 || *result > dataSize)
-                                newOffset = 0;
-                            else
-                                newOffset = *result;
-                        } else {
-                            runGoto = false;
-                        }
-                    }
+        ShortcutManager::addShortcut(this, CTRL + Keys::C, [this] {
+            const auto selection = this->getSelection();
+            copyBytes(selection);
+        });
 
-                    ImGui::EndTabItem();
-                }
-                if (ImGui::BeginTabItem("hex.builtin.view.hex_editor.goto.offset.current"_lang)) {
-                    ImGui::SetKeyboardFocusHere();
-                    runGoto = ImGui::InputText("##goto", this->m_gotoAddressInput.data(), this->m_gotoAddressInput.size(), ImGuiInputTextFlags_EnterReturnsTrue | ImGuiInputTextFlags_CallbackResize, ImGui::UpdateStringSizeCallback, &this->m_gotoAddressInput) || runGoto;
-                    if (runGoto) {
-                        MathEvaluator evaluator;
-                        auto result = evaluator.evaluate(this->m_gotoAddressInput);
-                        if (result) {
-                            i64 currSelectionOffset = std::min(this->m_memoryEditor.DataPreviewAddr, this->m_memoryEditor.DataPreviewAddrEnd);
+        ShortcutManager::addShortcut(this, CTRL + SHIFT + Keys::C, [this] {
+            const auto selection = this->getSelection();
+            copyString(selection);
+        });
 
-                            newOffset = 0;
-                            if (currSelectionOffset + *result < 0)
-                                newOffset = -currSelectionOffset;
-                            else if (currSelectionOffset + *result > dataSize)
-                                newOffset = dataSize - currSelectionOffset;
-                            else
-                                newOffset = *result;
+        ShortcutManager::addShortcut(this, CTRL + Keys::V, [this] {
+            const auto selection = this->getSelection();
+            pasteBytes(selection);
+        });
 
-                            newOffset = currSelectionOffset + *result + baseAddress;
-                        } else {
-                            runGoto = false;
-                        }
-                    }
-
-
-                    ImGui::EndTabItem();
-                }
-                if (ImGui::BeginTabItem("hex.builtin.view.hex_editor.goto.offset.end"_lang)) {
-                    ImGui::SetKeyboardFocusHere();
-                    runGoto = ImGui::InputText("##goto", this->m_gotoAddressInput.data(), this->m_gotoAddressInput.size(), ImGuiInputTextFlags_EnterReturnsTrue | ImGuiInputTextFlags_CallbackResize, ImGui::UpdateStringSizeCallback, &this->m_gotoAddressInput) || runGoto;
-                    if (runGoto) {
-                        MathEvaluator evaluator;
-                        auto result = evaluator.evaluate(this->m_gotoAddressInput);
-                        if (result) {
-                            if (*result < 0 || *result > dataSize)
-                                newOffset = 0;
-
-                            newOffset = (baseAddress + dataSize) - *result - 1;
-                        } else {
-                            runGoto = false;
-                        }
-                    }
-
-                    ImGui::EndTabItem();
-                }
-
-                if (ImGui::Button("hex.builtin.view.hex_editor.menu.file.goto"_lang)) {
-                    this->m_evaluateGoto = true;
-                }
-
-                if (runGoto) {
-                    this->m_gotoRequested = true;
-                    provider->setCurrentPage(std::floor(double(newOffset - baseAddress) / hex::prv::Provider::PageSize));
-                    ImHexApi::HexEditor::setSelection(newOffset, 1);
-                }
-
-                ImGui::EndTabBar();
-            }
-
-            if (ImGui::IsKeyDown(ImGui::GetKeyIndex(ImGuiKey_Escape)))
-                ImGui::CloseCurrentPopup();
-
-            ImGui::EndPopup();
-        }
-    }
-
-    void ViewHexEditor::drawEditPopup() {
-        auto provider      = ImHexApi::Provider::get();
-        bool providerValid = ImHexApi::Provider::isValid();
-        if (ImGui::MenuItem("hex.builtin.view.hex_editor.menu.edit.undo"_lang, "CTRL + Z", false, providerValid && provider->canUndo()))
-            provider->undo();
-        if (ImGui::MenuItem("hex.builtin.view.hex_editor.menu.edit.redo"_lang, "CTRL + Y", false, providerValid && provider->canRedo()))
-            provider->redo();
-
-        ImGui::Separator();
-
-        bool bytesSelected = this->m_memoryEditor.DataPreviewAddr != static_cast<size_t>(-1) && this->m_memoryEditor.DataPreviewAddrEnd != static_cast<size_t>(-1);
-
-        if (ImGui::MenuItem("hex.builtin.view.hex_editor.menu.edit.copy"_lang, "CTRL + C", false, bytesSelected))
-            this->copyBytes();
-
-        if (ImGui::BeginMenu("hex.builtin.view.hex_editor.menu.edit.copy_as"_lang, bytesSelected)) {
-            if (ImGui::MenuItem("hex.builtin.view.hex_editor.copy.hex"_lang, "CTRL + SHIFT + C"))
-                this->copyString();
-
-            ImGui::Separator();
-
-            for (const auto &[unlocalizedName, callback] : ContentRegistry::DataFormatter::getEntries()) {
-                if (ImGui::MenuItem(LangEntry(unlocalizedName))) {
-                    size_t start = std::min(this->m_memoryEditor.DataPreviewAddr, this->m_memoryEditor.DataPreviewAddrEnd);
-                    size_t end   = std::max(this->m_memoryEditor.DataPreviewAddr, this->m_memoryEditor.DataPreviewAddrEnd);
-
-                    size_t copySize = (end - start) + 1;
-
-                    ImGui::SetClipboardText(callback(provider, start + provider->getBaseAddress() + provider->getCurrentPageAddress(), copySize).c_str());
-                }
-            }
-
-            ImGui::EndMenu();
-        }
-
-        if (ImGui::MenuItem("hex.builtin.view.hex_editor.menu.edit.paste"_lang, "CTRL + V", false, bytesSelected))
-            this->pasteBytes();
-
-        if (ImGui::MenuItem("hex.builtin.view.hex_editor.menu.edit.select_all"_lang, "CTRL + A", false, providerValid))
-            ImHexApi::HexEditor::setSelection(provider->getBaseAddress(), provider->getActualSize());
-
-        ImGui::Separator();
-
-        if (ImGui::MenuItem("hex.builtin.view.hex_editor.menu.edit.bookmark"_lang, nullptr, false,
-                this->m_memoryEditor.DataPreviewAddr != static_cast<size_t>(-1) && this->m_memoryEditor.DataPreviewAddrEnd != static_cast<size_t>(-1)))
-        {
-            auto base = provider->getBaseAddress();
-
-            size_t start = base + std::min(this->m_memoryEditor.DataPreviewAddr, this->m_memoryEditor.DataPreviewAddrEnd);
-            size_t end   = base + std::max(this->m_memoryEditor.DataPreviewAddr, this->m_memoryEditor.DataPreviewAddrEnd);
-
-            ImHexApi::Bookmarks::add(start, end - start + 1, {}, {});
-        }
-
-        if (ImGui::MenuItem("hex.builtin.view.hex_editor.menu.edit.set_base"_lang, nullptr, false, providerValid && provider->isReadable())) {
-            this->m_baseAddress = provider->getBaseAddress();
-            ImHexApi::Tasks::doLater([] { ImGui::OpenPopup("hex.builtin.view.hex_editor.menu.edit.set_base"_lang); });
-        }
-
-        if (ImGui::MenuItem("hex.builtin.view.hex_editor.menu.edit.resize"_lang, nullptr, false, providerValid && provider->isResizable())) {
-            ImHexApi::Tasks::doLater([this] {
-                this->m_resizeSize = ImHexApi::Provider::get()->getActualSize();
-                ImGui::OpenPopup("hex.builtin.view.hex_editor.menu.edit.resize"_lang);
+        ShortcutManager::addShortcut(this, CTRL + Keys::O, [] {
+            fs::openFileBrowser(fs::DialogMode::Open, {}, [](const auto &path) {
+                EventManager::post<RequestOpenFile>(path);
             });
-        }
+        });
 
-        if (ImGui::MenuItem("hex.builtin.view.hex_editor.menu.edit.insert"_lang, nullptr, false, providerValid && provider->isResizable())) {
-            ImHexApi::Tasks::doLater([this] {
-                this->m_resizeSize = 0;
-                ImGui::OpenPopup("hex.builtin.view.hex_editor.menu.edit.insert"_lang);
-            });
-        }
+        ShortcutManager::addShortcut(this, CTRL + Keys::Z, [] {
+            if (ImHexApi::Provider::isValid())
+                ImHexApi::Provider::get()->undo();
+        });
+
+        ShortcutManager::addShortcut(this, CTRL + Keys::Y, [] {
+            if (ImHexApi::Provider::isValid())
+                ImHexApi::Provider::get()->redo();
+        });
     }
 
     void ViewHexEditor::registerEvents() {
-        EventManager::subscribe<RequestOpenFile>(this, [this](const auto &path) {
-            this->openFile(path);
-            this->getWindowOpenState() = true;
+        EventManager::subscribe<EventRegionSelected>(this, [this](auto) {
+            this->m_shouldModifyValue = true;
         });
 
         EventManager::subscribe<RequestSelectionChange>(this, [this](Region region) {
             auto provider = ImHexApi::Provider::get();
-            auto page     = provider->getPageOfAddress(region.address);
+            auto page     = provider->getPageOfAddress(region.getStartAddress());
 
             if (!page.has_value())
                 return;
 
             if (region.size != 0) {
                 provider->setCurrentPage(page.value());
-                u64 start = region.address - provider->getBaseAddress() - provider->getCurrentPageAddress();
-                this->m_memoryEditor.GotoAddrAndSelect(start, start + region.size - 1);
-            }
-
-            EventManager::post<EventRegionSelected>(Region { this->m_memoryEditor.DataPreviewAddr, (this->m_memoryEditor.DataPreviewAddrEnd - this->m_memoryEditor.DataPreviewAddr) + 1 });
-        });
-
-        EventManager::subscribe<EventProjectFileLoad>(this, []() {
-            EventManager::post<RequestOpenFile>(ProjectFile::getFilePath());
-        });
-
-        EventManager::subscribe<EventWindowClosing>(this, [](GLFWwindow *window) {
-            if (ProjectFile::hasUnsavedChanges()) {
-                glfwSetWindowShouldClose(window, GLFW_FALSE);
-                ImHexApi::Tasks::doLater([] { ImGui::OpenPopup("hex.builtin.view.hex_editor.exit_application.title"_lang); });
+                region.address -= provider->getBaseAddress() + provider->getCurrentPageAddress();
+                this->setSelection(region);
+                this->jumpToSelection();
             }
         });
 
-        EventManager::subscribe<RequestOpenWindow>(this, [this](const std::string &name) {
-            if (name == "Create File") {
-                fs::openFileBrowser(fs::DialogMode::Save, {}, [this](const auto &path) {
-                    fs::File file(path, fs::File::Mode::Create);
-
-                    if (!file.isValid()) {
-                        View::showErrorPopup("hex.builtin.view.hex_editor.error.create"_lang);
-                        return;
-                    }
-
-                    file.setSize(1);
-
-                    EventManager::post<RequestOpenFile>(path);
-                    this->getWindowOpenState() = true;
-                });
-            } else if (name == "Open File") {
-                fs::openFileBrowser(fs::DialogMode::Open, {}, [this](const auto &path) {
-                    EventManager::post<RequestOpenFile>(path);
-                    this->getWindowOpenState() = true;
-                });
-            } else if (name == "Open Project") {
-                fs::openFileBrowser(fs::DialogMode::Open, { {"Project File", "hexproj"} },
-                    [this](const auto &path) {
-                        ProjectFile::load(path);
-                        this->getWindowOpenState() = true;
-                    });
-            }
+        EventManager::subscribe<QuerySelection>(this, [this](auto &region) {
+            region = this->getSelection();
         });
 
         EventManager::subscribe<EventSettingsChanged>(this, [this] {
@@ -936,131 +1157,47 @@ namespace hex::plugin::builtin {
             }
 
             {
-                auto columnCount = ContentRegistry::Settings::getSetting("hex.builtin.setting.hex_editor", "hex.builtin.setting.hex_editor.bytes_per_row");
+                auto bytesPerRow = ContentRegistry::Settings::getSetting("hex.builtin.setting.hex_editor", "hex.builtin.setting.hex_editor.bytes_per_row");
 
-                if (columnCount.is_number())
-                    this->m_memoryEditor.Cols = static_cast<int>(columnCount);
-            }
-
-            {
-                auto hexii = ContentRegistry::Settings::getSetting("hex.builtin.setting.hex_editor", "hex.builtin.setting.hex_editor.hexii");
-
-                if (hexii.is_number())
-                    this->m_memoryEditor.OptShowHexII = static_cast<int>(hexii);
+                if (bytesPerRow.is_number())
+                    this->m_bytesPerRow = static_cast<int>(bytesPerRow);
             }
 
             {
                 auto ascii = ContentRegistry::Settings::getSetting("hex.builtin.setting.hex_editor", "hex.builtin.setting.hex_editor.ascii");
 
                 if (ascii.is_number())
-                    this->m_memoryEditor.OptShowAscii = static_cast<int>(ascii);
-            }
-
-            {
-                auto advancedDecoding = ContentRegistry::Settings::getSetting("hex.builtin.setting.hex_editor", "hex.builtin.setting.hex_editor.advanced_decoding");
-
-                if (advancedDecoding.is_number())
-                    this->m_advancedDecodingEnabled = static_cast<int>(advancedDecoding);
+                    this->m_showAscii = static_cast<int>(ascii);
             }
 
             {
                 auto greyOutZeros = ContentRegistry::Settings::getSetting("hex.builtin.setting.hex_editor", "hex.builtin.setting.hex_editor.grey_zeros");
 
                 if (greyOutZeros.is_number())
-                    this->m_memoryEditor.OptGreyOutZeroes = static_cast<int>(greyOutZeros);
+                    this->m_grayOutZero = static_cast<int>(greyOutZeros);
             }
 
             {
                 auto upperCaseHex = ContentRegistry::Settings::getSetting("hex.builtin.setting.hex_editor", "hex.builtin.setting.hex_editor.uppercase_hex");
 
                 if (upperCaseHex.is_number())
-                    this->m_memoryEditor.OptUpperCaseHex = static_cast<int>(upperCaseHex);
+                    this->m_upperCaseHex = static_cast<int>(upperCaseHex);
             }
 
             {
-                auto showExtraInfo = ContentRegistry::Settings::getSetting("hex.builtin.setting.hex_editor", "hex.builtin.setting.hex_editor.extra_info");
+                auto &visualizers = ContentRegistry::HexEditor::impl::getVisualizers();
+                auto selectedVisualizer = ContentRegistry::Settings::getSetting("hex.builtin.setting.hex_editor", "hex.builtin.setting.hex_editor.visualizer");
 
-                if (showExtraInfo.is_number())
-                    this->m_memoryEditor.OptShowExtraInfo = static_cast<int>(showExtraInfo);
+                if (selectedVisualizer.is_string() && visualizers.contains(selectedVisualizer))
+                    this->m_currDataVisualizer = visualizers[selectedVisualizer];
+                else
+                    this->m_currDataVisualizer = visualizers["hex.builtin.visualizer.hexadecimal.8bit"];
             }
-        });
 
-        EventManager::subscribe<QuerySelection>(this, [this](auto &region) {
-            u64 address = std::min(this->m_memoryEditor.DataPreviewAddr, this->m_memoryEditor.DataPreviewAddrEnd);
-            size_t size = std::abs(i64(this->m_memoryEditor.DataPreviewAddrEnd) - i64(this->m_memoryEditor.DataPreviewAddr)) + 1;
-
-            region = Region { address, size };
-        });
-
-        EventManager::subscribe<EventHighlightingChanged>(this, [this] {
-            this->m_highlights.clear();
-        });
-
-        EventManager::subscribe<EventProviderChanged>(this, [](auto, auto) {
-            EventManager::post<EventHighlightingChanged>();
-        });
-    }
-
-    void ViewHexEditor::registerShortcuts() {
-        ShortcutManager::addGlobalShortcut(CTRL + Keys::S, [] {
-            save();
-        });
-
-        ShortcutManager::addGlobalShortcut(CTRL + SHIFT + Keys::S, [] {
-            saveAs();
-        });
-
-
-        ShortcutManager::addShortcut(this, CTRL + Keys::Z, [] {
-            if (ImHexApi::Provider::isValid())
-                ImHexApi::Provider::get()->undo();
-        });
-
-        ShortcutManager::addShortcut(this, CTRL + Keys::Y, [] {
-            if (ImHexApi::Provider::isValid())
-                ImHexApi::Provider::get()->redo();
-        });
-
-
-        ShortcutManager::addShortcut(this, CTRL + Keys::F, [this] {
-            this->m_searchRequested = true;
-            ImGui::OpenPopupInWindow(View::toWindowName("hex.builtin.view.hex_editor.name").c_str(),
-                "hex.builtin.view.hex_editor.menu.file.search"_lang);
-        });
-
-        ShortcutManager::addShortcut(this, CTRL + Keys::G, [this] {
-            this->m_gotoRequested = true;
-            ImGui::OpenPopupInWindow(View::toWindowName("hex.builtin.view.hex_editor.name").c_str(),
-                "hex.builtin.view.hex_editor.menu.file.goto"_lang);
-        });
-
-        ShortcutManager::addShortcut(this, CTRL + Keys::O, [] {
-            fs::openFileBrowser(fs::DialogMode::Open, {}, [](const auto &path) {
-                EventManager::post<RequestOpenFile>(path);
-            });
-        });
-
-
-        ShortcutManager::addShortcut(this, CTRL + Keys::C, [this] {
-            this->copyBytes();
-        });
-
-        ShortcutManager::addShortcut(this, CTRL + SHIFT + Keys::C, [this] {
-            this->copyString();
-        });
-
-        ShortcutManager::addShortcut(this, CTRL + Keys::V, [this] {
-            this->pasteBytes();
-        });
-
-        ShortcutManager::addShortcut(this, CTRL + Keys::A, [] {
-            auto provider = ImHexApi::Provider::get();
-            ImHexApi::HexEditor::setSelection(provider->getBaseAddress(), provider->getActualSize());
         });
     }
 
     void ViewHexEditor::registerMenuItems() {
-
         /* Basic operations */
 
         ContentRegistry::Interface::addMenuItem("hex.builtin.menu.file", 1100, [&] {
@@ -1129,14 +1266,14 @@ namespace hex::plugin::builtin {
 
                 View::showFileChooserPopup(paths, { {"Thingy Table File", "tbl"} },
                     [this](const auto &path) {
-                        this->m_currEncodingFile = EncodingFile(EncodingFile::Type::Thingy, path);
+                        this->m_currCustomEncoding = EncodingFile(EncodingFile::Type::Thingy, path);
                     });
             }
         });
 
 
         /* Import / Export */
-        ContentRegistry::Interface::addMenuItem("hex.builtin.menu.file", 1300, [&, this] {
+        ContentRegistry::Interface::addMenuItem("hex.builtin.menu.file", 1300, [&] {
             auto provider      = ImHexApi::Provider::get();
             bool providerValid = ImHexApi::Provider::isValid();
 
@@ -1144,34 +1281,42 @@ namespace hex::plugin::builtin {
             if (ImGui::BeginMenu("hex.builtin.view.hex_editor.menu.file.import"_lang)) {
                 if (ImGui::MenuItem("hex.builtin.view.hex_editor.menu.file.import.base64"_lang)) {
 
-                    fs::openFileBrowser(fs::DialogMode::Open, {}, [this](const auto &path) {
-                        fs::File file(path, fs::File::Mode::Read);
-                        if (!file.isValid()) {
+                    fs::openFileBrowser(fs::DialogMode::Open, {}, [](const auto &path) {
+                        fs::File inputFile(path, fs::File::Mode::Read);
+                        if (!inputFile.isValid()) {
                             View::showErrorPopup("hex.builtin.view.hex_editor.error.open"_lang);
                             return;
                         }
 
-                        auto base64 = file.readBytes();
+                        auto base64 = inputFile.readBytes();
 
                         if (!base64.empty()) {
-                            this->m_dataToSave = crypt::decode64(base64);
+                            auto data = crypt::decode64(base64);
 
-                            if (this->m_dataToSave.empty())
+                            if (data.empty())
                                 View::showErrorPopup("hex.builtin.view.hex_editor.base64.import_error"_lang);
-                            else
-                                ImGui::OpenPopup("hex.builtin.view.hex_editor.save_data"_lang);
-                            this->getWindowOpenState() = true;
-                        } else View::showErrorPopup("hex.builtin.view.hex_editor.file_open_error"_lang);
+                            else {
+                                fs::openFileBrowser(fs::DialogMode::Save, {}, [&data](const std::fs::path &path) {
+                                    fs::File outputFile(path, fs::File::Mode::Create);
+
+                                    if (!outputFile.isValid())
+                                        View::showErrorPopup("hex.builtin.view.hex_editor.base64.import_error"_lang);
+
+                                    outputFile.write(data);
+                                });
+                            }
+                        } else {
+                            View::showErrorPopup("hex.builtin.view.hex_editor.file_open_error"_lang);
+                        }
                     });
                 }
 
                 ImGui::Separator();
 
-                if (ImGui::MenuItem("hex.builtin.view.hex_editor.menu.file.import.ips"_lang, nullptr, false, !this->m_processingImportExport)) {
+                if (ImGui::MenuItem("hex.builtin.view.hex_editor.menu.file.import.ips"_lang, nullptr, false)) {
 
-                    fs::openFileBrowser(fs::DialogMode::Open, {}, [this](const auto &path) {
-                        this->m_processingImportExport = true;
-                        std::thread([this, path] {
+                    fs::openFileBrowser(fs::DialogMode::Open, {}, [](const auto &path) {
+                        std::thread([path] {
                             auto task = ImHexApi::Tasks::createTask("hex.builtin.view.hex_editor.processing", 0);
 
                             auto patchData = fs::File(path, fs::File::Mode::Read).readBytes();
@@ -1189,17 +1334,13 @@ namespace hex::plugin::builtin {
                             }
 
                             provider->createUndoPoint();
-                            this->m_processingImportExport = false;
                         }).detach();
-
-                        this->getWindowOpenState() = true;
                     });
                 }
 
-                if (ImGui::MenuItem("hex.builtin.view.hex_editor.menu.file.import.ips32"_lang, nullptr, false, !this->m_processingImportExport)) {
-                    fs::openFileBrowser(fs::DialogMode::Open, {}, [this](const auto &path) {
-                        this->m_processingImportExport = true;
-                        std::thread([this, path] {
+                if (ImGui::MenuItem("hex.builtin.view.hex_editor.menu.file.import.ips32"_lang, nullptr, false)) {
+                    fs::openFileBrowser(fs::DialogMode::Open, {}, [](const auto &path) {
+                        std::thread([path] {
                             auto task = ImHexApi::Tasks::createTask("hex.builtin.view.hex_editor.processing", 0);
 
                             auto patchData = fs::File(path, fs::File::Mode::Read).readBytes();
@@ -1217,17 +1358,8 @@ namespace hex::plugin::builtin {
                             }
 
                             provider->createUndoPoint();
-                            this->m_processingImportExport = false;
                         }).detach();
-
-                        this->getWindowOpenState() = true;
                     });
-                }
-
-                if (ImGui::MenuItem("hex.builtin.view.hex_editor.menu.file.import.script"_lang)) {
-                    this->m_loaderScriptFilePath.clear();
-                    this->m_loaderScriptScriptPath.clear();
-                    ImHexApi::Tasks::doLater([] { ImGui::OpenPopup("hex.builtin.view.hex_editor.script.title"_lang); });
                 }
 
                 ImGui::EndMenu();
@@ -1236,7 +1368,7 @@ namespace hex::plugin::builtin {
 
             /* Export */
             if (ImGui::BeginMenu("hex.builtin.view.hex_editor.menu.file.export"_lang, providerValid && provider->isWritable())) {
-                if (ImGui::MenuItem("hex.builtin.view.hex_editor.menu.file.export.ips"_lang, nullptr, false, !this->m_processingImportExport)) {
+                if (ImGui::MenuItem("hex.builtin.view.hex_editor.menu.file.export.ips"_lang, nullptr, false)) {
                     Patches patches = provider->getPatches();
                     if (!patches.contains(0x00454F45) && patches.contains(0x00454F46)) {
                         u8 value = 0;
@@ -1244,28 +1376,26 @@ namespace hex::plugin::builtin {
                         patches[0x00454F45] = value;
                     }
 
-                    this->m_processingImportExport = true;
-                    std::thread([this, patches] {
+                    std::thread([patches] {
                         auto task = ImHexApi::Tasks::createTask("hex.builtin.view.hex_editor.processing", 0);
 
-                        this->m_dataToSave             = generateIPSPatch(patches);
-                        this->m_processingImportExport = false;
+                        auto data = generateIPSPatch(patches);
 
-                        ImHexApi::Tasks::doLater([this] {
-                            fs::openFileBrowser(fs::DialogMode::Save, {}, [this](const auto &path) {
+                        ImHexApi::Tasks::doLater([data] {
+                            fs::openFileBrowser(fs::DialogMode::Save, {}, [&data](const auto &path) {
                                 auto file = fs::File(path, fs::File::Mode::Create);
                                 if (!file.isValid()) {
                                     View::showErrorPopup("hex.builtin.view.hex_editor.error.create"_lang);
                                     return;
                                 }
 
-                                file.write(this->m_dataToSave);
+                                file.write(data);
                             });
                         });
                     }).detach();
                 }
 
-                if (ImGui::MenuItem("hex.builtin.view.hex_editor.menu.file.export.ips32"_lang, nullptr, false, !this->m_processingImportExport)) {
+                if (ImGui::MenuItem("hex.builtin.view.hex_editor.menu.file.export.ips32"_lang, nullptr, false)) {
                     Patches patches = provider->getPatches();
                     if (!patches.contains(0x00454F45) && patches.contains(0x45454F46)) {
                         u8 value = 0;
@@ -1273,22 +1403,20 @@ namespace hex::plugin::builtin {
                         patches[0x45454F45] = value;
                     }
 
-                    this->m_processingImportExport = true;
-                    std::thread([this, patches] {
+                    std::thread([patches] {
                         auto task = ImHexApi::Tasks::createTask("hex.builtin.view.hex_editor.processing", 0);
 
-                        this->m_dataToSave             = generateIPS32Patch(patches);
-                        this->m_processingImportExport = false;
+                        auto data = generateIPS32Patch(patches);
 
-                        ImHexApi::Tasks::doLater([this] {
-                            fs::openFileBrowser(fs::DialogMode::Save, {}, [this](const auto &path) {
+                        ImHexApi::Tasks::doLater([data] {
+                            fs::openFileBrowser(fs::DialogMode::Save, {}, [&data](const auto &path) {
                                 auto file = fs::File(path, fs::File::Mode::Create);
                                 if (!file.isValid()) {
                                     View::showErrorPopup("hex.builtin.view.hex_editor.error.create"_lang);
                                     return;
                                 }
 
-                                file.write(this->m_dataToSave);
+                                file.write(data);
                             });
                         });
                     }).detach();
@@ -1304,23 +1432,83 @@ namespace hex::plugin::builtin {
             bool providerValid = ImHexApi::Provider::isValid();
 
             if (ImGui::MenuItem("hex.builtin.view.hex_editor.menu.file.search"_lang, "CTRL + F", false, providerValid)) {
-                this->getWindowOpenState() = true;
-                this->m_searchRequested    = true;
-                ImGui::OpenPopupInWindow(View::toWindowName("hex.builtin.view.hex_editor.name").c_str(), "hex.builtin.view.hex_editor.menu.file.search"_lang);
+                this->openPopup<PopupFind>();
             }
 
             if (ImGui::MenuItem("hex.builtin.view.hex_editor.menu.file.goto"_lang, "CTRL + G", false, providerValid)) {
-                this->getWindowOpenState() = true;
-                this->m_gotoRequested      = true;
-                ImGui::OpenPopupInWindow(View::toWindowName("hex.builtin.view.hex_editor.name").c_str(), "hex.builtin.view.hex_editor.menu.file.goto"_lang);
+                this->openPopup<PopupGoto>();
             }
         });
 
 
         /* Edit menu */
         ContentRegistry::Interface::addMenuItem("hex.builtin.menu.edit", 1000, [&, this] {
-            this->drawEditPopup();
+            this->drawEditMenu();
         });
+    }
+
+    void ViewHexEditor::drawEditMenu() {
+        auto provider      = ImHexApi::Provider::get();
+        bool providerValid = ImHexApi::Provider::isValid();
+        if (ImGui::MenuItem("hex.builtin.view.hex_editor.menu.edit.undo"_lang, "CTRL + Z", false, providerValid && provider->canUndo()))
+            provider->undo();
+        if (ImGui::MenuItem("hex.builtin.view.hex_editor.menu.edit.redo"_lang, "CTRL + Y", false, providerValid && provider->canRedo()))
+            provider->redo();
+
+        ImGui::Separator();
+
+        const bool bytesSelected = this->isSelectionValid();
+        const auto selection = this->getSelection();
+
+        if (ImGui::MenuItem("hex.builtin.view.hex_editor.menu.edit.copy"_lang, "CTRL + C", false, bytesSelected))
+            copyBytes(selection);
+
+        if (ImGui::BeginMenu("hex.builtin.view.hex_editor.menu.edit.copy_as"_lang, bytesSelected)) {
+            if (ImGui::MenuItem("hex.builtin.view.hex_editor.copy.hex"_lang, "CTRL + SHIFT + C"))
+                copyString(selection);
+
+            ImGui::Separator();
+
+            for (const auto &[unlocalizedName, callback] : ContentRegistry::DataFormatter::getEntries()) {
+                if (ImGui::MenuItem(LangEntry(unlocalizedName))) {
+                    ImGui::SetClipboardText(
+                        callback(
+                            provider,
+                            selection.getStartAddress() + provider->getBaseAddress() + provider->getCurrentPageAddress(),
+                            selection.size
+                        ).c_str()
+                    );
+                }
+            }
+
+            ImGui::EndMenu();
+        }
+
+        if (ImGui::MenuItem("hex.builtin.view.hex_editor.menu.edit.paste"_lang, "CTRL + V", false, bytesSelected))
+            pasteBytes(selection);
+
+        if (ImGui::MenuItem("hex.builtin.view.hex_editor.menu.edit.select_all"_lang, "CTRL + A", false, providerValid))
+            ImHexApi::HexEditor::setSelection(provider->getBaseAddress(), provider->getActualSize());
+
+        ImGui::Separator();
+
+        if (ImGui::MenuItem("hex.builtin.view.hex_editor.menu.edit.bookmark"_lang, nullptr, false, bytesSelected)) {
+            auto base = provider->getBaseAddress();
+
+            ImHexApi::Bookmarks::add(base + selection.getStartAddress(), selection.getEndAddress(), {}, {});
+        }
+
+        if (ImGui::MenuItem("hex.builtin.view.hex_editor.menu.edit.set_base"_lang, nullptr, false, providerValid && provider->isReadable())) {
+            this->openPopup<PopupBaseAddress>(provider->getBaseAddress());
+        }
+
+        if (ImGui::MenuItem("hex.builtin.view.hex_editor.menu.edit.resize"_lang, nullptr, false, providerValid && provider->isResizable())) {
+            this->openPopup<PopupResize>(provider->getActualSize());
+        }
+
+        if (ImGui::MenuItem("hex.builtin.view.hex_editor.menu.edit.insert"_lang, nullptr, false, providerValid && provider->isResizable())) {
+            this->openPopup<PopupInsert>(this->getSelection().getStartAddress(), 0x00);
+        }
     }
 
 }
